feat(frontend): introduce esbuild + split app.js into domain ES modules
Replaces the 2,249-line monolithic agentception/static/app.js with 11
domain-specific ES modules under agentception/static/js/, bundled by
esbuild into the same output path so no template changes are required.

- package.json: adds esbuild devDependency with build/watch scripts
- static/js/nav.js: projectSwitcher
- static/js/overview.js: pipelineDashboard, agentCard, phaseSwitcher,
  pipelineControl, sweepControl, waveControl, scalingAdvisor,
  prViolations, staleClaimCard, issueCard
- static/js/agents.js: agentsPage, missionControl
- static/js/telemetry.js: telemetryDash
- static/js/dag.js: dagVisualization
- static/js/config.js: configPanel
- static/js/roles.js: roleDetail, rolesEditor
- static/js/brain_dump.js: brainDump
- static/js/transcripts.js: transcriptBrowser, transcriptDetail
- static/js/templates.js: exportPanel, importPanel, envSandbox
- static/js/api.js: apiEndpoint
- static/js/app.js: entry point — imports all modules, assigns to window

All 25 factory functions are kept as globals via Object.assign(window, ...)
so every x-data="functionName()" template attribute is unchanged.
