+++ .gitignore
@@ -1,3 +1,6 @@
+# Node / JS tooling
+node_modules/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[codz]
