+++ agentception/static/app.js
@@ -1,2249 +1,10 @@
-/**
- * AgentCeption — Alpine.js component library
- *
- * All Alpine component factory functions live here so templates stay logic-free.
- * Templates call these by name:  x-data="componentName(serverRenderedArgs)"
- *
- * Jinja2 data is injected at the call-site in the template attribute, never
- * inside this file.  This keeps the file static, cacheable, and free of
- * server-side template rendering bugs caused by mismatched quote styles.
- *
- * Function index:
- *   projectSwitcher()                         — nav project dropdown
- *   pipelineDashboard(initial)                — overview SSE board
- *   agentCard()                               — expandable agent row with live transcript
- *   phaseSwitcher(label, labels, pinned)      — phase dropdown
- *   pipelineControl(paused)                   — pause/resume toggle
- *   waveControl()                             — start-wave button
- *   scalingAdvisor(initial)                   — scaling recommendation banner
- *   prViolations(initial)                     — out-of-order PR banner
- *   staleClaimCard(claim)                     — stale-claim clear action
- *   telemetryDash()                           — telemetry D3 dashboard tab switching
- *   configPanel(initial)                      — pipeline config editor
- *   dagVisualization()                        — dependency DAG D3 graph
- *   spawnForm()                               — manual spawn form
- *   exportPanel()                             — template export
- *   importPanel()                             — template import
- *   transcriptBrowser(q, role, status)        — transcript list filters/sort
- *   transcriptDetail()                        — in-conversation search
- */
-
-'use strict';
-
-// ---------------------------------------------------------------------------
-// Navigation — project switcher
-// ---------------------------------------------------------------------------
-
-/**
- * Fetches the pipeline-config and renders a project <select> in the nav bar.
- * Hidden via x-show when no projects are configured.
- */
-function projectSwitcher() {
-  return {
-    projects: [],
-    activeProject: null,
-
-    async load() {
-      try {
-        const res = await fetch('/api/config');
-        if (!res.ok) return;
-        const cfg = await res.json();
-        this.projects = cfg.projects || [];
-        this.activeProject = cfg.active_project || null;
-      } catch (_) { /* network error — silently suppress */ }
-    },
-
-    async switchProject(name) {
-      if (!name || name === this.activeProject) return;
-      try {
-        const res = await fetch('/api/config/switch-project', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ project_name: name }),
-        });
-        if (res.ok) window.location.reload();
-      } catch (_) { /* silently suppress */ }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — pipeline dashboard (SSE-driven)
-// ---------------------------------------------------------------------------
-
-/**
- * Receives a PipelineState snapshot from the server as the initial value,
- * then subscribes to GET /events (SSE) for live updates.
- * Never polls — fully event-driven once connected.
- *
- * @param {object} initial - Server-rendered PipelineState snapshot.
- */
-function pipelineDashboard(initial) {
-  return {
-    state: initial,
-    connected: false,
-    boardLoading: false,
-    _es: null,
-    // GH_BASE_URL is provided via data-gh-base-url on the root element so no
-    // inline <script> tag is needed.  Set in connect() once $el is available.
-    GH_BASE_URL: '',
-
-    // Global kill modal — lives here so SSE re-renders never reset it.
-    killModal: { show: false, agentId: null, issueNumber: null, killing: false },
-
-    connect() {
-      this.GH_BASE_URL = this.$el.dataset.ghBaseUrl ?? '';
-      this._es = new EventSource('/events');
-
-      this._es.onopen = () => { this.connected = true; };
-
-      this._es.onmessage = (e) => {
-        try {
-          this.state = JSON.parse(e.data);
-          this.connected = true;
-        } catch (_) {
-          // Malformed JSON — keep stale state.
-        }
-      };
-
-      this._es.onerror = () => {
-        this.connected = false;
-        // EventSource auto-reconnects; just update the indicator.
-      };
-    },
-
-    /** Open the kill confirmation modal for a specific agent. */
-    openKillModal(agentId, issueNumber) {
-      this.killModal = { show: true, agentId, issueNumber: issueNumber ?? null, killing: false };
-    },
-
-    /** Close the kill modal without acting. */
-    closeKillModal() {
-      this.killModal.show = false;
-    },
-
-    /** Perform the kill: DELETE worktree via API, then close modal. */
-    async confirmKill() {
-      const id = this.killModal.agentId;
-      if (!id) return;
-      this.killModal.killing = true;
-      try {
-        await fetch(`/api/control/kill/${encodeURIComponent(id)}`, { method: 'POST' });
-      } catch (e) {
-        console.error('Kill failed', e);
-      } finally {
-        this.killModal = { show: false, agentId: null, issueNumber: null, killing: false };
-      }
-    },
-
-    /** Format a UNIX timestamp as a human-readable relative time string. */
-    relativeTime(ts) {
-      if (!ts) return '—';
-      const secs = Math.floor(Date.now() / 1000 - ts);
-      if (secs < 5)    return 'just now';
-      if (secs < 60)   return secs + 's ago';
-      if (secs < 3600) return Math.floor(secs / 60) + 'm ago';
-      return Math.floor(secs / 3600) + 'h ago';
-    },
-
-    /** Find the first live agent working on a given issue number (for board cross-reference). */
-    agentForIssue(issueNumber) {
-      if (!this.state || !this.state.agents) return null;
-      return this.state.agents.find(a => a.issue_number === issueNumber) || null;
-    },
-
-    /** Sum of message_count across all currently tracked agents. */
-    totalMsgCount() {
-      if (!this.state || !this.state.agents) return 0;
-      return this.state.agents.reduce((sum, a) => sum + (a.message_count || 0), 0);
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — expandable agent card with live transcript
-// ---------------------------------------------------------------------------
-
-/**
- * Powers each agent row in the Agent Tree.
- *
- * The card manages its own expansion state and the on-demand transcript fetch.
- * Reactive agent data (status, issue_number, etc.) lives in the parent SSE
- * scope and is passed in via x-effect so re-fetching triggers automatically
- * when message_count changes.
- *
- * Usage in template:
- *   <div x-data="agentCard()" x-effect="checkRefresh(agent, agent.message_count)">
- */
-function agentCard() {
-  return {
-    expanded:         false,
-    transcript:       [],
-    transcriptLoading: false,
-    transcriptError:  null,
-    _prevMsgCount:    0,
-
-    /** Toggle expand/collapse; fetch transcript on first open. */
-    toggle(agentId) {
-      this.expanded = !this.expanded;
-      if (this.expanded && this.transcript.length === 0) {
-        this.fetchTranscript(agentId);
-      }
-    },
-
-    /**
-     * Fetch the agent's transcript from the REST endpoint.
-     * Called on expand and when message_count changes while expanded.
-     */
-    async fetchTranscript(agentId) {
-      this.transcriptLoading = true;
-      this.transcriptError   = null;
-      try {
-        const res = await fetch(`/api/agents/${encodeURIComponent(agentId)}/transcript`);
-        if (!res.ok) throw new Error(`HTTP ${res.status}`);
-        const data = await res.json();
-        // API may return an array directly or {messages:[]} — handle both.
-        this.transcript = Array.isArray(data) ? data : (data.messages || []);
-        this._scrollFeed();
-      } catch (err) {
-        this.transcriptError = err.message || 'Failed to load transcript';
-      } finally {
-        this.transcriptLoading = false;
-      }
-    },
-
-    /**
-     * Called from x-effect="checkRefresh(agent, agent.message_count)".
-     * The explicit agent.message_count arg makes Alpine track it as a dependency
-     * so this re-runs on every SSE tick where the count changed.
-     */
-    checkRefresh(agent, msgCount) {
-      const count = msgCount || 0;
-      if (this.expanded && count !== this._prevMsgCount) {
-        this.fetchTranscript(agent.id);
-      }
-      this._prevMsgCount = count;
-    },
-
-    /** Scroll the transcript feed to the latest message. */
-    _scrollFeed() {
-      this.$nextTick(() => {
-        const feed = this.$el.querySelector('.transcript-feed');
-        if (feed) feed.scrollTop = feed.scrollHeight;
-      });
-    },
-
-    /** Format a kebab-case role name as Title Case words. */
-    formatRole(role) {
-      return (role || '').split('-')
-        .map(w => w.charAt(0).toUpperCase() + w.slice(1))
-        .join(' ');
-    },
-
-    /** True for statuses that mean the agent is actively working. */
-    isActive(status) {
-      return status === 'implementing' || status === 'reviewing';
-    },
-
-    /** Clip long text for compact transcript preview. */
-    truncate(text, maxLen) {
-      if (!text) return '';
-      return text.length > maxLen ? text.slice(0, maxLen) + '…' : text;
-    },
-
-    /**
-     * Parse a COGNITIVE_ARCH string into display-ready fields.
-     * Format: "figure_id:skill1:skill2"
-     * Returns an object with emoji, displayName, archetype, archetypeKey, skillDomains.
-     */
-    parseArch(archStr) {
-      if (!archStr) return { emoji: '🤖', displayName: '', archetype: '', archetypeKey: 'default', skillDomains: [], figureId: null };
-
-      const FIGURE_MAP = {
-        steve_jobs:         { name: 'Steve Jobs',          emoji: '🔮', archetype: 'The Visionary',   key: 'visionary' },
-        satya_nadella:      { name: 'Satya Nadella',       emoji: '🏛️', archetype: 'The Architect',   key: 'architect' },
-        jeff_bezos:         { name: 'Jeff Bezos',          emoji: '⚙️', archetype: 'The Operator',    key: 'operator' },
-        werner_vogels:      { name: 'Werner Vogels',       emoji: '🏛️', archetype: 'The Architect',   key: 'architect' },
-        linus_torvalds:     { name: 'Linus Torvalds',      emoji: '⚡', archetype: 'The Hacker',      key: 'hacker' },
-        margaret_hamilton:  { name: 'Margaret Hamilton',   emoji: '🛡️', archetype: 'The Guardian',    key: 'guardian' },
-        bjarne_stroustrup:  { name: 'Bjarne Stroustrup',   emoji: '⚡', archetype: 'The Hacker',      key: 'hacker' },
-        martin_fowler:      { name: 'Martin Fowler',       emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        kent_beck:          { name: 'Kent Beck',           emoji: '🔧', archetype: 'The Pragmatist',  key: 'pragmatist' },
-        yann_lecun:         { name: 'Yann LeCun',          emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        andrej_karpathy:    { name: 'Andrej Karpathy',     emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        turing:             { name: 'Alan Turing',         emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        hopper:             { name: 'Grace Hopper',        emoji: '🧑‍🏫', archetype: 'The Mentor',   key: 'mentor' },
-        dijkstra:           { name: 'Edsger Dijkstra',     emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        knuth:              { name: 'Donald Knuth',        emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        ritchie:            { name: 'Dennis Ritchie',      emoji: '⚡', archetype: 'The Hacker',      key: 'hacker' },
-        guido_van_rossum:   { name: 'Guido van Rossum',   emoji: '🔧', archetype: 'The Pragmatist',  key: 'pragmatist' },
-        mccarthy:           { name: 'John McCarthy',       emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        lovelace:           { name: 'Ada Lovelace',        emoji: '🔮', archetype: 'The Visionary',   key: 'visionary' },
-        von_neumann:        { name: 'John von Neumann',    emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        shannon:            { name: 'Claude Shannon',      emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        feynman:            { name: 'Richard Feynman',     emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        einstein:           { name: 'Albert Einstein',     emoji: '🔮', archetype: 'The Visionary',   key: 'visionary' },
-        hamming:            { name: 'Richard Hamming',     emoji: '📚', archetype: 'The Scholar',     key: 'scholar' },
-        bruce_schneier:     { name: 'Bruce Schneier',      emoji: '🛡️', archetype: 'The Guardian',    key: 'guardian' },
-      };
-
-      const ARCHETYPE_MAP = {
-        the_visionary:  { emoji: '🔮', name: 'The Visionary',  key: 'visionary' },
-        the_architect:  { emoji: '🏛️', name: 'The Architect',  key: 'architect' },
-        the_hacker:     { emoji: '⚡', name: 'The Hacker',     key: 'hacker' },
-        the_guardian:   { emoji: '🛡️', name: 'The Guardian',   key: 'guardian' },
-        the_scholar:    { emoji: '📚', name: 'The Scholar',    key: 'scholar' },
-        the_mentor:     { emoji: '🧑‍🏫', name: 'The Mentor',  key: 'mentor' },
-        the_operator:   { emoji: '⚙️', name: 'The Operator',   key: 'operator' },
-        the_pragmatist: { emoji: '🔧', name: 'The Pragmatist', key: 'pragmatist' },
-      };
-
-      const parts = archStr.split(':').map(p => p.trim()).filter(Boolean);
-      const first = parts[0];
-      const figData = FIGURE_MAP[first];
-
-      if (figData) {
-        return {
-          figureId: first,
-          displayName: figData.name,
-          emoji: figData.emoji,
-          archetype: figData.archetype,
-          archetypeKey: figData.key,
-          skillDomains: parts.slice(1),
-        };
-      }
-
-      // Check if first part is an archetype
-      const archData = ARCHETYPE_MAP[first];
-      if (archData) {
-        return {
-          figureId: null,
-          displayName: archData.name,
-          emoji: archData.emoji,
-          archetype: archData.name,
-          archetypeKey: archData.key,
-          skillDomains: parts.slice(1),
-        };
-      }
-
-      // Unknown — treat all parts as skill domains
-      return {
-        figureId: null,
-        displayName: parts.join(' · '),
-        emoji: '🤖',
-        archetype: '',
-        archetypeKey: 'default',
-        skillDomains: parts,
-      };
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — phase switcher dropdown
-// ---------------------------------------------------------------------------
-
-/**
- * Phase-label selector in the pipeline board header.
- * Dispatches 'phase-switching' so the board can show a loading spinner
- * before the fetch + page reload.
- *
- * @param {string|null} initialLabel  - Currently active label.
- * @param {string[]}    allLabels     - All configured phase labels.
- * @param {boolean}     initialPinned - Whether the label is manually pinned.
- */
-function phaseSwitcher(initialLabel, allLabels, initialPinned) {
-  return {
-    current: initialLabel,
-    labels: allLabels,
-    pinned: initialPinned,
-    open: false,
-
-    async selectLabel(label) {
-      this.open = false;
-      this.$dispatch('phase-switching');
-      try {
-        const res = await fetch('/api/control/active-label', {
-          method: 'PUT',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ label }),
-        });
-        if (res.ok) {
-          const data = await res.json();
-          this.current = data.label;
-          this.pinned = data.pinned;
-          window.location.reload();
-        }
-      } catch (_) {}
-    },
-
-    async selectAuto() {
-      this.open = false;
-      this.$dispatch('phase-switching');
-      try {
-        const res = await fetch('/api/control/active-label', { method: 'DELETE' });
-        if (res.ok) {
-          const data = await res.json();
-          this.current = data.label;
-          this.pinned = false;
-          window.location.reload();
-        }
-      } catch (_) {}
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — pipeline pause/resume control
-// ---------------------------------------------------------------------------
-
-/**
- * Toggle the background poller between paused and running.
- *
- * @param {boolean} initialPaused - Server-rendered current pause state.
- */
-function pipelineControl(initialPaused) {
-  return {
-    paused: initialPaused ?? false,
-
-    async toggle() {
-      const endpoint = this.paused ? '/api/control/resume' : '/api/control/pause';
-      try {
-        const res = await fetch(endpoint, { method: 'POST' });
-        if (res.ok) {
-          const data = await res.json();
-          this.paused = data.paused;
-        }
-      } catch (_) {
-        // Network error — leave state unchanged.
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — sweep stale branches / claims
-// ---------------------------------------------------------------------------
-
-/**
- * Calls POST /api/control/sweep to delete stale branches and clear orphan
- * agent:wip labels in one shot.  Lives inside the "Stale" summary-item so
- * it shares the same Alpine scope as the count badge.
- */
-function sweepControl() {
-  return {
-    sweeping: false,
-
-    async sweep() {
-      if (!confirm('Delete all stale agent branches and clear orphan agent:wip labels?')) return;
-      this.sweeping = true;
-      try {
-        const resp = await fetch('/api/control/sweep', { method: 'POST' });
-        const data = await resp.json();
-        if (!resp.ok) {
-          alert(`Sweep failed: ${data.detail ?? resp.status}`);
-          return;
-        }
-        const b = data.deleted_branches?.length ?? 0;
-        const w = data.cleared_wip_labels?.length ?? 0;
-        const e = data.errors?.length ?? 0;
-        const msg = [`Swept: ${b} branch${b !== 1 ? 'es' : ''} deleted, ${w} label${w !== 1 ? 's' : ''} cleared`];
-        if (e > 0) msg.push(`(${e} error${e !== 1 ? 's' : ''} — check logs)`);
-        alert(msg.join(' '));
-      } catch (err) {
-        alert(`Network error: ${err.message}`);
-      } finally {
-        this.sweeping = false;
-      }
-    },
-  };
-}
-
-// Overview — wave control (start-wave button)
-// ---------------------------------------------------------------------------
-
-/**
- * Counts unclaimed board issues and fires POST /api/control/spawn-wave.
- * Data flows from the parent pipelineDashboard state via Alpine scope.
- */
-function waveControl() {
-  return {
-    launching: false,
-    waveResult: null,
-    waveError: null,
-
-    init() {
-      // Data flows from the parent pipelineDashboard scope via SSE — no polling needed here.
-    },
-
-    unclaimedCount() {
-      const issues = this.state?.board_issues;
-      if (!issues) return 0;
-      return issues.filter(i => !i.claimed).length;
-    },
-
-    claimedCount() {
-      const issues = this.state?.board_issues;
-      if (!issues) return 0;
-      return issues.filter(i => i.claimed).length;
-    },
-
-    async startWave() {
-      this.launching  = true;
-      this.waveResult = null;
-      this.waveError  = null;
-      try {
-        const resp = await fetch('/api/control/spawn-wave', { method: 'POST' });
-        const data = await resp.json();
-        if (!resp.ok) {
-          this.waveError = data.detail ?? `HTTP ${resp.status}`;
-        } else {
-          this.waveResult = data;
-        }
-      } catch (err) {
-        this.waveError = `Network error: ${err.message}`;
-      } finally {
-        this.launching = false;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — scaling advisor banner
-// ---------------------------------------------------------------------------
-
-/**
- * Shows a recommendation banner when the pool sizing is off.
- * "Apply" posts to the API; "Dismiss" hides locally for the session.
- *
- * @param {object|null} initial - Server-rendered ScalingAdvice object.
- */
-function scalingAdvisor(initial) {
-  return {
-    recommendation: initial ?? null,
-    dismissed: false,
-    applying: false,
-    applyError: null,
-
-    async apply() {
-      this.applying   = true;
-      this.applyError = null;
-      try {
-        const res = await fetch('/api/intelligence/scaling-advice/apply', { method: 'POST' });
-        if (res.ok) {
-          this.dismissed = true;
-        } else {
-          const data = await res.json().catch(() => ({}));
-          this.applyError = data.detail || `HTTP ${res.status}`;
-        }
-      } catch (err) {
-        this.applyError = err.message || 'Network error';
-      } finally {
-        this.applying = false;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — PR violations banner
-// ---------------------------------------------------------------------------
-
-/**
- * Displays out-of-order PR cards and lets the user close the PR from here.
- *
- * @param {object[]|null} initial - Server-rendered list of PRViolation objects.
- */
-function prViolations(initial) {
-  return {
-    violations: initial ?? [],
-    closing: null,
-
-    async closeViolation(prNumber) {
-      this.closing = prNumber;
-      try {
-        const res = await fetch(
-          `/api/intelligence/pr-violations/${prNumber}/close`,
-          { method: 'POST' },
-        );
-        if (res.ok) {
-          this.violations = this.violations.filter(v => v.pr_number !== prNumber);
-        }
-      } catch (_) {
-        // Network error — leave the card visible so the user can retry.
-      } finally {
-        this.closing = null;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — stale claim card
-// ---------------------------------------------------------------------------
-
-/**
- * Two-step confirm → clear flow for a single stale agent:wip claim.
- *
- * @param {object} claim - StaleClaim object from PipelineState.stale_claims.
- */
-function staleClaimCard(claim) {
-  return {
-    confirming: false,
-    clearing: false,
-    cleared: false,
-    error: null,
-
-    async clear(issueNumber) {
-      this.confirming = false;
-      this.clearing   = true;
-      this.error      = null;
-      try {
-        const res = await fetch(
-          `/api/intelligence/stale-claims/${issueNumber}/clear`,
-          { method: 'POST' },
-        );
-        if (res.ok) {
-          this.cleared = true;
-        } else {
-          const data = await res.json().catch(() => ({}));
-          this.error = data.detail || `HTTP ${res.status}`;
-        }
-      } catch (err) {
-        this.error = err.message || 'Network error';
-      } finally {
-        this.clearing = false;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Overview — issue card (Analyze button)
-// ---------------------------------------------------------------------------
-
-/**
- * Powers each issue card in the GitHub Board's open issues list.
- *
- * Keeps the inline fetch out of the template.  The analysisHtml is injected
- * via x-html after the POST returns so no HTMX processing is needed.
- *
- * @param {number} issueNumber - GitHub issue number for the API call.
- */
-function issueCard(issueNumber) {
-  return {
-    analysisHtml: '',
-    analyzing: false,
-    analyzeError: null,
-
-    async analyze() {
-      this.analyzing = true;
-      this.analyzeError = null;
-      try {
-        const r = await fetch(`/api/analyze/issue/${issueNumber}/partial`, { method: 'POST' });
-        this.analysisHtml = await r.text();
-      } catch (err) {
-        this.analyzeError = err.message;
-      } finally {
-        this.analyzing = false;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Telemetry — sparkline trend chart
-// ---------------------------------------------------------------------------
-
-/**
- * Draws a three-series canvas sparkline (issues, PRs, agents).
- * No external charting library required.
- *
- * @param {string[]} labels  - X-axis time labels.
- * @param {number[]} issues  - Issue counts per bucket.
- * @param {number[]} prs     - PR counts per bucket.
- * @param {number[]} agents  - Agent counts per bucket.
- */
-// ---------------------------------------------------------------------------
-// Telemetry — D3 dashboard tab controller
-// ---------------------------------------------------------------------------
-
-/**
- * telemetryDash()
- *
- * Minimal Alpine component that owns tab state for the telemetry D3 dashboard.
- * The actual D3 rendering lives in telemetry.js (loaded only on /telemetry).
- *
- * Tab IDs map to window.telemetry render functions:
- *   'Gantt' → window.telemetry.renderGantt
- *   'CostArea' → window.telemetry.renderCostArea
- *   etc.
- */
-function telemetryDash() {
-  return {
-    activeTab: 'Trend',
-
-    switchTab(tab) {
-      this.activeTab = tab;
-      this.$nextTick(() => {
-        if (window.telemetry && window.telemetry['render' + tab]) {
-          window.telemetry['render' + tab]();
-        }
-      });
-    },
-
-    init() {
-      this.$nextTick(() => {
-        if (window.telemetry) {
-          window.telemetry.renderTrend();
-        }
-      });
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Config — pipeline config editor
-// ---------------------------------------------------------------------------
-
-/**
- * Full config editor with drag-and-drop label reordering and save/toast.
- * Initial config is passed from the server; init() re-fetches to stay fresh.
- *
- * @param {object|null} initialConfig - Server-rendered PipelineConfig dict.
- */
-function configPanel() {
-  const defaults = {
-    max_eng_vps: 1,
-    max_qa_vps: 1,
-    pool_size_per_vp: 4,
-    active_labels_order: [],
-    ab_mode: { enabled: false, target_role: null, variant_a_file: null, variant_b_file: null },
-    projects: [],
-    active_project: null,
-  };
-  return {
-    // ── State ──────────────────────────────────────────────────────────────
-    config: { ...defaults },
-    lastSaved: null,
-    activeTab: 'allocation',
-    newLabel: '',
-    saving: false,
-    toast: { msg: '', cls: '' },
-    projectSaving: null,
-    savedAt: null,
-    _dragIdx: null,
-
-    // ── Computed capacity ──────────────────────────────────────────────────
-    get totalAgents() {
-      return (this.config.max_eng_vps + this.config.max_qa_vps) * this.config.pool_size_per_vp;
-    },
-    get engSlots() { return this.config.max_eng_vps * this.config.pool_size_per_vp; },
-    get qaSlots()  { return this.config.max_qa_vps  * this.config.pool_size_per_vp; },
-    get engPct()   { return this.totalAgents ? Math.round(this.engSlots / this.totalAgents * 100) : 50; },
-    get qaPct()    { return this.totalAgents ? Math.round(this.qaSlots  / this.totalAgents * 100) : 50; },
-
-    // ── Lifecycle ──────────────────────────────────────────────────────────
-    async init() {
-      // Hydrate from server-rendered data attribute first (zero flicker)
-      const raw = this.$el.dataset.config;
-      if (raw) {
-        try { this.config = { ...defaults, ...JSON.parse(raw) }; }
-        catch (_) { /* fall back to defaults */ }
-      }
-      // Re-fetch for freshness
-      try {
-        const r = await fetch('/api/config');
-        if (!r.ok) throw new Error('HTTP ' + r.status);
-        this.config = { ...defaults, ...await r.json() };
-      } catch (err) {
-        this._showToast('Failed to load config: ' + err.message, 'err');
-      }
-      this.lastSaved = JSON.parse(JSON.stringify(this.config));
-    },
-
-    // ── Tab navigation ─────────────────────────────────────────────────────
-    setTab(tab) { this.activeTab = tab; },
-
-    // ── Label editor ───────────────────────────────────────────────────────
-    addLabel() {
-      const label = this.newLabel.trim();
-      if (!label) return;
-      if (this.config.active_labels_order.includes(label)) {
-        this._showToast('Label already in list.', 'err');
-        return;
-      }
-      this.config.active_labels_order.push(label);
-      this.newLabel = '';
-    },
-
-    removeLabel(idx) { this.config.active_labels_order.splice(idx, 1); },
-
-    onDragStart(evt, idx) {
-      this._dragIdx = idx;
-      evt.dataTransfer.effectAllowed = 'move';
-      evt.dataTransfer.setData('text/plain', String(idx));
-    },
-
-    onDragOver(evt)  { evt.currentTarget.classList.add('drag-over'); },
-    onDragLeave(evt) { evt.currentTarget.classList.remove('drag-over'); },
-
-    onDrop(evt, targetIdx) {
-      evt.currentTarget.classList.remove('drag-over');
-      const srcIdx = this._dragIdx;
-      if (srcIdx === null || srcIdx === targetIdx) return;
-      const arr = this.config.active_labels_order;
-      const [item] = arr.splice(srcIdx, 1);
-      arr.splice(targetIdx, 0, item);
-      this._dragIdx = null;
-    },
-
-    onDragEnd() {
-      document.querySelectorAll('.cfg-label-item.drag-over').forEach(el => {
-        el.classList.remove('drag-over');
-      });
-      this._dragIdx = null;
-    },
-
-    // ── A/B mode ───────────────────────────────────────────────────────────
-    ensureAbMode() {
-      if (!this.config.ab_mode) {
-        this.config.ab_mode = { enabled: false, target_role: null, variant_a_file: null, variant_b_file: null };
-      }
-    },
-
-    // ── Projects ───────────────────────────────────────────────────────────
-    async switchProject(name) {
-      this.projectSaving = name;
-      try {
-        const r = await fetch('/api/config/switch-project', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ project_name: name }),
-        });
-        if (!r.ok) {
-          const detail = await r.json().then(d => d.detail || r.statusText).catch(() => r.statusText);
-          throw new Error(detail);
-        }
-        const updated = await r.json();
-        this.config.active_project = updated.active_project;
-        this.lastSaved = JSON.parse(JSON.stringify(this.config));
-        this._showToast('✅ Switched to ' + name, 'ok');
-      } catch (err) {
-        this._showToast('Switch failed: ' + err.message, 'err');
-      } finally {
-        this.projectSaving = null;
-      }
-    },
-
-    // ── Save / Cancel ──────────────────────────────────────────────────────
-    cancel() {
-      if (!this.lastSaved) return;
-      this.config = JSON.parse(JSON.stringify(this.lastSaved));
-      this._showToast('Changes discarded.', '');
-    },
-
-    get isDirty() {
-      return JSON.stringify(this.config) !== JSON.stringify(this.lastSaved);
-    },
-
-    async save() {
-      this.saving = true;
-      this._showToast('', '');
-      try {
-        const r = await fetch('/api/config', {
-          method: 'PUT',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(this.config),
-        });
-        if (!r.ok) {
-          const detail = await r.json().then(d => d.detail || r.statusText).catch(() => r.statusText);
-          throw new Error(detail);
-        }
-        this.config = { ...defaults, ...await r.json() };
-        this.lastSaved = JSON.parse(JSON.stringify(this.config));
-        this.savedAt = new Date().toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-        this._showToast('✅ Config saved.', 'ok');
-      } catch (err) {
-        this._showToast('Save failed: ' + err.message, 'err');
-      } finally {
-        this.saving = false;
-      }
-    },
-
-    // ── Helpers ────────────────────────────────────────────────────────────
-    _showToast(msg, cls) { this.toast = { msg, cls }; },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Spawn — Mission Control orchestration dashboard
-// ---------------------------------------------------------------------------
-
-/**
- * missionControl()
- *
- * Drives the spawn Mission Control page. Three modes:
- *   - 'single'      → POST /api/control/spawn (one issue + role)
- *   - 'wave'        → POST /api/control/spawn-wave (all unclaimed issues)
- *   - 'coordinator' → POST /api/control/spawn-coordinator (brain dump)
- *
- * Data is hydrated from data-* attributes on $el inside init() so that
- * server-side JSON (which uses double-quotes) never collides with the HTML
- * x-data="..." attribute boundary.
- */
-function missionControl() {
-  return {
-    // ── Mode ──────────────────────────────────────────────────
-    mode: 'single',
-
-    // ── Single-agent state ────────────────────────────────────
-    selectedIssue: null,   // full issue object
-    selectedRole: 'python-developer',
-    searchQ: '',
-
-    // ── Wave state ────────────────────────────────────────────
-    waveRole: 'python-developer',
-
-    // ── Coordinator state ─────────────────────────────────────
-    brainDump: '',
-    labelPrefix: '',
-
-    // ── Shared ────────────────────────────────────────────────
-    loading: false,
-    result: null,
-    resultMode: null,   // which mode produced the result
-    formError: null,
-
-    // ── Issue data (hydrated from data-* in init) ─────────────
-    issues: [],
-    activeLabel: '',
-    unclaimedCount: 0,
-
-    init() {
-      try {
-        this.issues = JSON.parse(this.$el.dataset.issues || '[]');
-      } catch (_) {
-        this.issues = [];
-      }
-      try {
-        this.activeLabel = JSON.parse(this.$el.dataset.activeLabel || '""');
-      } catch (_) {
-        this.activeLabel = '';
-      }
-      this.unclaimedCount = parseInt(this.$el.dataset.unclaimedCount || '0', 10);
-    },
-
-    // ── Computed ──────────────────────────────────────────────
-
-    /** Issues visible in the board after filtering (single mode). */
-    get filteredIssues() {
-      const q = this.searchQ.trim().toLowerCase();
-      if (!q) return this.issues;
-      return this.issues.filter(iss =>
-        String(iss.number).includes(q) ||
-        (iss.title || '').toLowerCase().includes(q) ||
-        (iss.phase_label || '').toLowerCase().includes(q)
-      );
-    },
-
-    /** Count of unclaimed issues (used in wave mode button label). */
-    get liveUnclaimedCount() {
-      return this.issues.filter(i => !i.claimed).length;
-    },
-
-    // ── Actions ───────────────────────────────────────────────
-
-    selectIssue(iss) {
-      if (iss.claimed) return;
-      this.selectedIssue = (this.selectedIssue?.number === iss.number) ? null : iss;
-    },
-
-    selectRole(slug) {
-      this.selectedRole = slug;
-    },
-
-    clearResult() {
-      this.result    = null;
-      this.formError = null;
-      this.resultMode = null;
-    },
-
-    /** POST /api/control/spawn — spawn a single engineer agent. */
-    async submitSingle() {
-      if (!this.selectedIssue) return;
-      this.clearResult();
-      this.loading = true;
-      try {
-        const resp = await fetch('/api/control/spawn', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            issue_number: this.selectedIssue.number,
-            role: this.selectedRole,
-          }),
-        });
-        const data = await resp.json();
-        if (!resp.ok) {
-          this.formError = data.detail ?? `HTTP ${resp.status}`;
-        } else {
-          this.result     = data;
-          this.resultMode = 'single';
-          // Mark the issue claimed in local list so the board updates instantly.
-          this.issues = this.issues.map(i =>
-            i.number === this.selectedIssue.number ? { ...i, claimed: true } : i
-          );
-        }
-      } catch (err) {
-        this.formError = `Network error: ${err.message}`;
-      } finally {
-        this.loading = false;
-      }
-    },
-
-    /** POST /api/control/spawn-wave — spawn all unclaimed issues at once. */
-    async submitWave() {
-      this.clearResult();
-      this.loading = true;
-      try {
-        const resp = await fetch(
-          `/api/control/spawn-wave?role=${encodeURIComponent(this.waveRole)}`,
-          { method: 'POST' }
-        );
-        const data = await resp.json();
-        if (!resp.ok) {
-          this.formError = data.detail ?? `HTTP ${resp.status}`;
-        } else {
-          this.result     = data;
-          this.resultMode = 'wave';
-          // Refresh local issue list to reflect newly claimed issues.
-          const spawnedNums = new Set((data.spawned || []).map(s => s.spawned));
-          this.issues = this.issues.map(i =>
-            spawnedNums.has(i.number) ? { ...i, claimed: true } : i
-          );
-        }
-      } catch (err) {
-        this.formError = `Network error: ${err.message}`;
-      } finally {
-        this.loading = false;
-      }
-    },
-
-    /** POST /api/control/spawn-coordinator — launch a coordinator brain-dump. */
-    async submitCoordinator() {
-      if (!this.brainDump.trim()) return;
-      this.clearResult();
-      this.loading = true;
-      try {
-        const resp = await fetch('/api/control/spawn-coordinator', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            brain_dump: this.brainDump,
-            label_prefix: this.labelPrefix,
-          }),
-        });
-        const data = await resp.json();
-        if (!resp.ok) {
-          this.formError = data.detail ?? `HTTP ${resp.status}`;
-        } else {
-          this.result     = data;
-          this.resultMode = 'coordinator';
-        }
-      } catch (err) {
-        this.formError = `Network error: ${err.message}`;
-      } finally {
-        this.loading = false;
-      }
-    },
-
-    /** Called by HTMX after the issue list partial swaps in. */
-    onIssuesRefreshed(newIssues) {
-      this.issues = newIssues;
-      // Preserve selection if the selected issue still exists.
-      if (this.selectedIssue) {
-        const still = this.issues.find(i => i.number === this.selectedIssue.number);
-        this.selectedIssue = still ?? null;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Templates — export / import panels
-// ---------------------------------------------------------------------------
-
-/** Role template export panel. */
-function exportPanel() {
-  return {
-    name: '',
-    version: '',
-    busy: false,
-    toast: { msg: '', cls: '' },
-
-    async doExport() {
-      if (!this.name || !this.version) return;
-      this.busy = true;
-      this._toast('', '');
-      try {
-        const r = await fetch('/api/templates/export', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ name: this.name, version: this.version }),
-        });
-        if (!r.ok) {
-          const d = await r.json().catch(() => ({ detail: r.statusText }));
-          throw new Error(d.detail || r.statusText);
-        }
-        const blob = await r.blob();
-        const cd   = r.headers.get('Content-Disposition') || '';
-        const match = cd.match(/filename="([^"]+)"/);
-        const filename = match ? match[1] : 'template.tar.gz';
-        const url = URL.createObjectURL(blob);
-        const a   = document.createElement('a');
-        a.href = url; a.download = filename; a.click();
-        URL.revokeObjectURL(url);
-        this._toast('✅ Exported — page will refresh to show the new entry.', 'ok');
-        setTimeout(() => location.reload(), 1500);
-      } catch (err) {
-        this._toast('Export failed: ' + err.message, 'err');
-      } finally {
-        this.busy = false;
-      }
-    },
-
-    _toast(msg, cls) { this.toast = { msg, cls }; },
-  };
-}
-
-/** Role template import panel. */
-function importPanel() {
-  return {
-    file: null,
-    targetRepo: '',
-    busy: false,
-    toast: { msg: '', cls: '' },
-    conflicts: [],
-
-    onFileChange(evt) { this.file = evt.target.files[0] || null; },
-
-    async doImport() {
-      if (!this.file || !this.targetRepo) return;
-      this.busy = true;
-      this.conflicts = [];
-      this._toast('', '');
-      try {
-        const fd  = new FormData();
-        fd.append('file', this.file);
-        const url = '/api/templates/import?target_repo=' + encodeURIComponent(this.targetRepo);
-        const r   = await fetch(url, { method: 'POST', body: fd });
-        if (!r.ok) {
-          const d = await r.json().catch(() => ({ detail: r.statusText }));
-          throw new Error(d.detail || r.statusText);
-        }
-        const result  = await r.json();
-        this.conflicts = (result.conflicts || []).filter(c => c.exists);
-        this._toast('✅ Imported ' + result.extracted.length + ' file(s) (' + result.name + ' v' + result.version + ').', 'ok');
-      } catch (err) {
-        this._toast('Import failed: ' + err.message, 'err');
-      } finally {
-        this.busy = false;
-      }
-    },
-
-    _toast(msg, cls) { this.toast = { msg, cls }; },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Transcripts — list browser
-// ---------------------------------------------------------------------------
-
-/**
- * Client-side filter and sort for the transcript browser table.
- * Server pre-filters via query params; this handles live typing without
- * a round-trip.
- *
- * @param {string} initialQ      - Server-applied ?q= value.
- * @param {string} initialRole   - Server-applied ?role= value.
- * @param {string} initialStatus - Server-applied ?status= value.
- */
-function transcriptBrowser(initialQ, initialRole, initialStatus) {
-  return {
-    q: initialQ || '',
-    filterRole: initialRole || '',
-    filterStatus: initialStatus || '',
-    sort: 'mtime',
-
-    applyFilters() {
-      const qLow = this.q.toLowerCase();
-      document.querySelectorAll('[data-transcript-row]').forEach(el => {
-        const preview = (el.dataset.preview || '').toLowerCase();
-        const role    = el.dataset.role    || '';
-        const status  = el.dataset.status  || '';
-        const matchQ  = !qLow || preview.includes(qLow);
-        const matchR  = !this.filterRole   || role   === this.filterRole;
-        const matchS  = !this.filterStatus || status === this.filterStatus;
-        el.style.display = (matchQ && matchR && matchS) ? '' : 'none';
-      });
-      this.applySortDom();
-    },
-
-    applySortDom() {
-      const tbody = document.getElementById('transcript-tbody');
-      if (!tbody) return;
-      const rows = Array.from(tbody.querySelectorAll('[data-transcript-row]'));
-      rows.sort((a, b) => {
-        switch (this.sort) {
-          case 'messages': return parseInt(b.dataset.messages) - parseInt(a.dataset.messages);
-          case 'words':    return parseInt(b.dataset.words)    - parseInt(a.dataset.words);
-          case 'subagents':return parseInt(b.dataset.subagents)- parseInt(a.dataset.subagents);
-          default:         return parseFloat(b.dataset.mtime)  - parseFloat(a.dataset.mtime);
-        }
-      });
-      rows.forEach(r => tbody.appendChild(r));
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Transcript detail — in-conversation search + highlight
-// ---------------------------------------------------------------------------
-
-/**
- * Powers the in-thread search field on the transcript detail page.
- * Filters messages client-side via Alpine x-show and injects <mark> tags
- * for matched text via x-html.
- */
-function transcriptDetail() {
-  return {
-    search: '',
-
-    highlight(text) {
-      if (!this.search) return this._esc(text);
-      const escaped = this.search.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-      const re = new RegExp(escaped, 'gi');
-      return this._esc(text).replace(re, m => `<mark class="hl">${m}</mark>`);
-    },
-
-    /** HTML-escape plain text before injecting highlight marks. */
-    _esc(text) {
-      return String(text)
-        .replace(/&/g, '&amp;')
-        .replace(/</g, '&lt;')
-        .replace(/>/g, '&gt;')
-        .replace(/"/g, '&quot;');
-    },
-  };
-}
-
-// ═══ Cognitive Architecture Studio ══════════════════════════════════════════
-//
-// roleDetail(slug, fileExists, personasJson)
-//   Alpine component for the center panel (HTMX partial: _role_detail.html).
-//   Manages: tab state, persona selection + apply-to-composer, composer form,
-//   and the "View / Edit Prompt" button that explicitly wakes the Monaco editor.
-//
-//   Performance note: init() does NOT auto-dispatch `role-load`. The freeze
-//   on role click was caused by Monaco.setValue() running synchronously on
-//   every HTMX swap. Now the user must click "View / Edit Prompt" to load
-//   the editor — making it deliberate and non-blocking.
-//
-// rolesEditor()
-//   Alpine component for the right panel (Monaco editor, roles.html).
-//   Initialises Monaco lazily on the first `role-load` window event.
-//   Uses a debounced ResizeObserver instead of automaticLayout to avoid
-//   continuous layout thrashing.
-// ─────────────────────────────────────────────────────────────────────────────
-
-function roleDetail(slug, fileExists, personas) {
-  return {
-    slug,
-    fileExists,
-    personas,
-    activeTab: 'personas',
-    selectedPersonaId: null,
-    figure: '',
-    atomOverrides: {},
-    skills: [],
-    copied: false,
-
-    init() {
-      // No auto-dispatch here — deliberate. See performance note above.
-    },
-
-    // Called by the "View / Edit Prompt" button in _role_detail.html.
-    loadInEditor() {
-      window.dispatchEvent(new CustomEvent('role-load', { detail: { slug: this.slug } }));
-    },
-
-    get archPreview() {
-      const parts = [];
-      if (this.figure) parts.push(this.figure);
-      for (const s of this.skills) parts.push(s);
-      return parts.length ? `COGNITIVE_ARCH=${parts.join(':')}` : '(select a figure or skill)';
-    },
-
-    applyPersona(id) {
-      const persona = this.personas.find(p => p.id === id);
-      if (!persona) return;
-      this.figure = id;
-      this.atomOverrides = persona.overrides ? { ...persona.overrides } : {};
-      this.skills = [];
-      this.activeTab = 'composer';
-    },
-
-    async copyArchString() {
-      try {
-        await navigator.clipboard.writeText(this.archPreview);
-      } catch (_) {
-        // Clipboard API may be unavailable in non-secure context.
-      }
-      this.copied = true;
-      setTimeout(() => { this.copied = false; }, 1500);
-    },
-
-    resetComposer() {
-      this.figure = '';
-      this.atomOverrides = {};
-      this.skills = [];
-    },
-  };
-}
-
-function rolesEditor() {
-  return {
-    editor: null,
-    _monacoBooted: false,  // true once _bootMonaco() has been called
-    _monacoReady: false,   // true once monaco.editor.create() has returned
-    _pendingSlug: null,
-    currentSlug: null,
-    currentPath: null,
-    status: '',
-    statusClass: '',
-    breadcrumb: '← click "View / Edit Prompt" to open a file',
-    canSave: false,
-    canDiff: false,
-    diffVisible: false,
-    diffTitle: '',
-    diffLines: [],
-    diffCommitReady: false,
-    diffCommitting: false,
-
-    // No x-init. This component is fully inert until loadRole() is called.
-    // This means 0 bytes of Monaco JS are downloaded on page load.
-
-    async loadRole(slug) {
-      // Triggered by @role-load.window when user clicks "View / Edit Prompt".
-      // First call boots Monaco by injecting the loader script dynamically,
-      // then loads the file. Subsequent calls skip straight to _doLoad().
-      if (!this._monacoReady) {
-        this._pendingSlug = slug;
-        if (!this._monacoBooted) {
-          this._monacoBooted = true;
-          this.setStatus('Loading editor…', '');
-          this._injectMonaco();
-        }
-        return;
-      }
-      await this._doLoad(slug);
-    },
-
-    _injectMonaco() {
-      // Dynamically inject the Monaco AMD loader — only on first use.
-      // Keeps 1.5 MB+ of Monaco JS off the page until the user asks for it.
-      const MONACO_VERSION = '0.52.0';
-      const CDN = `https://cdn.jsdelivr.net/npm/monaco-editor@${MONACO_VERSION}/min/vs`;
-
-      const script = document.createElement('script');
-      script.src = `${CDN}/loader.js`;
-      script.onload = () => {
-        require.config({ paths: { vs: CDN } });
-        require(['vs/editor/editor.main'], () => {
-          if (this.editor) return; // Guard against double-init.
-
-          if (this.$refs.editorPlaceholder) {
-            this.$refs.editorPlaceholder.style.display = 'none';
-          }
-
-          this.editor = monaco.editor.create(this.$refs.editorContainer, {
-            value: '',
-            language: 'markdown',
-            theme: 'vs-dark',
-            automaticLayout: false,
-            minimap: { enabled: false },
-            wordWrap: 'on',
-            scrollBeyondLastLine: false,
-            readOnly: true,
-          });
-
-          // Debounced ResizeObserver — layout() at most once per 100 ms.
-          let _t = null;
-          new ResizeObserver(() => {
-            clearTimeout(_t);
-            _t = setTimeout(() => this.editor && this.editor.layout(), 100);
-          }).observe(this.$refs.editorContainer);
-
-          this._monacoReady = true;
-
-          if (this._pendingSlug) {
-            this._doLoad(this._pendingSlug);
-            this._pendingSlug = null;
-          }
-        });
-      };
-      document.head.appendChild(script);
-    },
-
-    async _doLoad(slug) {
-      if (!this.editor) return;
-      const path = `.cursor/roles/${slug}.md`;
-      this.setStatus(`Loading ${path}…`, '');
-      try {
-        const r = await fetch(`/api/roles/${encodeURIComponent(slug)}`);
-        if (!r.ok) throw new Error(`HTTP ${r.status}`);
-        const data = await r.json();
-
-        // Use requestAnimationFrame so the browser can paint the loading status
-        // before Monaco's synchronous setValue() blocks the thread.
-        requestAnimationFrame(() => {
-          this.editor.setValue(data.content);
-          this.editor.updateOptions({ readOnly: false });
-          this.editor.layout();
-          this.currentSlug = slug;
-          this.currentPath = path;
-          this.breadcrumb = path;
-          this.canSave = true;
-          this.canDiff = true;
-          const msg = data.meta.last_commit_message || '(uncommitted)';
-          this.setStatus(`${data.meta.line_count} lines · ${msg}`, 'ok');
-        });
-      } catch (err) {
-        this.setStatus(`Failed to load: ${err.message}`, 'err');
-      }
-    },
-
-    async saveRole() {
-      if (!this.editor || !this.currentSlug) return;
-      this.canSave = false;
-      this.setStatus('Saving…', '');
-      try {
-        const r = await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}`, {
-          method: 'PUT',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ content: this.editor.getValue() }),
-        });
-        if (!r.ok) throw new Error(`HTTP ${r.status}`);
-        const data = await r.json();
-        this.canSave = true;
-        const lines = data.diff ? data.diff.split('\n').length : 0;
-        this.setStatus(`✅ Saved — ${lines} line(s) changed`, 'ok');
-      } catch (err) {
-        this.canSave = true;
-        this.setStatus(`Save failed: ${err.message}`, 'err');
-      }
-    },
-
-    async previewDiff() {
-      if (!this.editor || !this.currentSlug) return;
-      this.diffTitle = `Diff — ${this.currentPath}`;
-      this.diffLines = [];
-      this.diffCommitReady = false;
-      this.diffVisible = true;
-      try {
-        const r = await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}/diff`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ content: this.editor.getValue() }),
-        });
-        if (!r.ok) throw new Error(`HTTP ${r.status}`);
-        const data = await r.json();
-        this.diffLines = this._parseDiff(data.diff);
-        this.diffCommitReady = true;
-      } catch (err) {
-        this.diffLines = [{ cls: 'cas-diff-empty', text: `Failed: ${err.message}` }];
-      }
-    },
-
-    _parseDiff(diff) {
-      if (!diff || !diff.trim()) {
-        return [{ cls: 'cas-diff-empty', text: 'No changes — content is identical to HEAD.' }];
-      }
-      return diff.split('\n').map(line => {
-        let cls = 'cas-diff-line';
-        if (line.startsWith('+') && !line.startsWith('+++')) cls += ' cas-diff-line--added';
-        else if (line.startsWith('-') && !line.startsWith('---')) cls += ' cas-diff-line--removed';
-        else if (line.startsWith('@@')) cls += ' cas-diff-line--hunk';
-        return { cls, text: line };
-      });
-    },
-
-    async saveAndCommit() {
-      if (!this.editor || !this.currentSlug) return;
-      this.diffCommitting = true;
-      try {
-        const r = await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}/commit`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ content: this.editor.getValue() }),
-        });
-        if (!r.ok) throw new Error(`HTTP ${r.status}`);
-        const data = await r.json();
-        this.diffVisible = false;
-        this.setStatus(`✅ Committed — SHA ${data.commit_sha.slice(0, 8)} · ${data.message}`, 'ok');
-      } catch (err) {
-        this.setStatus(`Commit failed: ${err.message}`, 'err');
-      } finally {
-        this.diffCommitting = false;
-      }
-    },
-
-    setStatus(msg, cls = '') {
-      this.status = msg;
-      this.statusClass = cls;
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Brain Dump page
-// ---------------------------------------------------------------------------
-
-/**
- * Powers the Brain Dump form — text input → loading animation → done panel.
- *
- * Static data (loading messages) is read from data attributes on the root
- * element so they live in Python/Jinja, not in this file.
- * Seeds are rendered server-side by Jinja; each button passes its text via
- * $el.dataset.text so no seed array is needed here.
- *
- * HTMX refreshes the recent-runs sidebar after a successful submit:
- *   htmx.trigger('#bd-recent-runs', 'refresh')
- */
-function brainDump() {
-  return {
-    step: 'input',
-    funnelStage: 0,
-    text: '',
-    labelPrefix: '',
-    showOptions: false,
-    focused: false,
-    submitting: false,
-    errorMsg: '',
-    result: {},
-    loadingMsg: '',
-    _loadingMsgs: [],
-    _loadingTimer: null,
-
-    init() {
-      // Loading messages are defined in Python, injected via data attribute.
-      this._loadingMsgs = JSON.parse(this.$el.dataset.loadingMsgs || '[]');
-      this.loadingMsg = this._loadingMsgs[0] ?? '';
-    },
-
-    get lineCount() {
-      return this.text.split('\n').filter(l => l.trim()).length;
-    },
-
-    stageClass(n) {
-      if (this.funnelStage > n) return 'bd-funnel-stage--done';
-      if (this.funnelStage === n && this.step !== 'input') return 'bd-funnel-stage--active bd-funnel-stage--pulse';
-      if (n === 0 && this.step === 'input') return 'bd-funnel-stage--active';
-      return '';
-    },
-
-    autoGrow(el) {
-      el.style.height = 'auto';
-      el.style.height = Math.min(el.scrollHeight, 600) + 'px';
-    },
-
-    async pasteClipboard() {
-      try {
-        const t = await navigator.clipboard.readText();
-        this.text = (this.text ? this.text + '\n' : '') + t;
-        await this.$nextTick();
-        this.autoGrow(this.$refs.textarea);
-      } catch (_) {}
-    },
-
-    // Called from Jinja-rendered seed buttons: @click="appendSeed($el.dataset.text)"
-    appendSeed(txt) {
-      this.text = (this.text.trim() ? this.text.trim() + '\n' : '') + txt;
-      this.$nextTick(() => this.autoGrow(this.$refs.textarea));
-    },
-
-    _startLoadingAnimation() {
-      let i = 0;
-      this.funnelStage = 1;
-      this.loadingMsg = this._loadingMsgs[0] ?? '';
-      this._loadingTimer = setInterval(() => {
-        i = (i + 1) % this._loadingMsgs.length;
-        this.loadingMsg = this._loadingMsgs[i] ?? '';
-        this.funnelStage = Math.min(i + 1, 5);
-      }, 700);
-    },
-
-    _stopLoadingAnimation(finalStage) {
-      if (this._loadingTimer) { clearInterval(this._loadingTimer); this._loadingTimer = null; }
-      this.funnelStage = finalStage;
-    },
-
-    async submit() {
-      const trimmed = this.text.trim();
-      if (!trimmed) return;
-      this.submitting = true;
-      this.errorMsg = '';
-      this.step = 'loading';
-      this._startLoadingAnimation();
-      try {
-        const resp = await fetch('/api/control/spawn-coordinator', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ brain_dump: trimmed, label_prefix: this.labelPrefix.trim() }),
-        });
-        if (!resp.ok) {
-          const body = await resp.json().catch(() => ({}));
-          throw new Error(body.detail || `HTTP ${resp.status}`);
-        }
-        this.result = await resp.json();
-        this._stopLoadingAnimation(5);
-        this.step = 'done';
-        // Refresh the recent-runs sidebar via HTMX so the new run appears.
-        const sidebar = document.getElementById('bd-recent-runs');
-        if (sidebar && typeof htmx !== 'undefined') htmx.trigger(sidebar, 'refresh');
-      } catch (err) {
-        this._stopLoadingAnimation(0);
-        this.errorMsg = err.message;
-        this.step = 'input';
-      } finally {
-        this.submitting = false;
-      }
-    },
-
-    reset() {
-      this._stopLoadingAnimation(0);
-      this.step = 'input';
-      this.funnelStage = 0;
-      this.text = '';
-      this.labelPrefix = '';
-      this.showOptions = false;
-      this.errorMsg = '';
-      this.result = {};
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// API Reference — per-endpoint execution
-// ---------------------------------------------------------------------------
-
-/**
- * Powers each endpoint card on /api.
- *
- * Data is read from data-* attributes on the root element so all static
- * values (method, path template, param specs, body field definitions) live
- * in Python/Jinja rather than in this file.
- *
- * data-method        — HTTP verb (GET, POST, …)
- * data-path          — path template, e.g. /agents/{agent_id}
- * data-query-params  — JSON array of {name, required, description} objects
- * data-body-fields   — JSON array of {name, type, required, default} objects
- */
-function apiEndpoint() {
-  return {
-    open: false,
-    tryOpen: false,
-    method: 'GET',
-    pathTemplate: '/',
-    pathParamNames: [],   // e.g. ['agent_id']
-    pathParamValues: {},  // {agent_id: ''}
-    queryParams: [],      // [{name, required, description}]
-    queryValues: {},      // {name: ''}
-    bodyFields: [],
-    bodyJson: '',
-    response: null,
-    sending: false,
-
-    init() {
-      this.method       = this.$el.dataset.method || 'GET';
-      this.pathTemplate = this.$el.dataset.path   || '/';
-
-      // Query params
-      try { this.queryParams = JSON.parse(this.$el.dataset.queryParams || '[]'); }
-      catch { this.queryParams = []; }
-      for (const p of this.queryParams) this.queryValues[p.name] = '';
-
-      // Body fields → JSON skeleton
-      try { this.bodyFields = JSON.parse(this.$el.dataset.bodyFields || '[]'); }
-      catch { this.bodyFields = []; }
-      if (this.bodyFields.length) {
-        const skeleton = {};
-        for (const f of this.bodyFields) {
-          if (f.default !== null && f.default !== undefined && f.default !== '') {
-            skeleton[f.name] = f.default;
-          } else if (f.type === 'integer' || f.type === 'number') {
-            skeleton[f.name] = 0;
-          } else if (f.type === 'boolean') {
-            skeleton[f.name] = false;
-          } else if (f.type && f.type.startsWith('array')) {
-            skeleton[f.name] = [];
-          } else {
-            skeleton[f.name] = '';
-          }
-        }
-        this.bodyJson = JSON.stringify(skeleton, null, 2);
-      }
-
-      // Path params — extract {name} tokens from template
-      const matches = this.pathTemplate.match(/\{(\w+)\}/g) || [];
-      this.pathParamNames = matches.map(m => m.slice(1, -1));
-      for (const n of this.pathParamNames) this.pathParamValues[n] = '';
-    },
-
-    get resolvedPath() {
-      let url = this.pathTemplate;
-      for (const [k, v] of Object.entries(this.pathParamValues)) {
-        url = url.replace(`{${k}}`, v ? encodeURIComponent(v) : `{${k}}`);
-      }
-      const qs = Object.entries(this.queryValues)
-        .filter(([, v]) => v !== '')
-        .map(([k, v]) => `${encodeURIComponent(k)}=${encodeURIComponent(v)}`);
-      if (qs.length) url += '?' + qs.join('&');
-      return url;
-    },
-
-    // Build a curl command that mirrors the current form state.
-    get curlCommand() {
-      const base = window.location.origin;
-      let cmd = `curl -X ${this.method} '${base}${this.resolvedPath}'`;
-      const hasBody = !['GET', 'HEAD'].includes(this.method) && this.bodyJson.trim();
-      if (hasBody) {
-        // Escape single quotes inside the JSON for a POSIX shell literal.
-        const escaped = this.bodyJson.replace(/'/g, "'\\''");
-        cmd += ` \\\n  -H 'Content-Type: application/json' \\\n  -d '${escaped}'`;
-      }
-      return cmd;
-    },
-
-    formatSize(bytes) {
-      if (bytes < 1024) return `${bytes} B`;
-      if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-      return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-    },
-
-    async send() {
-      this.sending = true;
-      this.response = null;
-      const t0 = performance.now();
-      try {
-        const opts = { method: this.method, headers: {} };
-        if (!['GET', 'HEAD', 'DELETE'].includes(this.method) && this.bodyJson.trim()) {
-          opts.headers['Content-Type'] = 'application/json';
-          opts.body = this.bodyJson;
-        }
-        const resp = await fetch(this.resolvedPath, opts);
-        const elapsed = Math.round(performance.now() - t0);
-        const text = await resp.text();
-
-        // Collect all response headers.
-        const headers = [];
-        resp.headers.forEach((value, name) => headers.push({ name, value }));
-        const contentType = resp.headers.get('content-type') || '';
-        const size = new TextEncoder().encode(text).length;
-
-        // Pretty-print JSON bodies; leave everything else as-is.
-        let body = text;
-        try { body = JSON.stringify(JSON.parse(text), null, 2); } catch { /* not JSON */ }
-
-        this.response = { status: resp.status, ok: resp.ok, body, elapsed, size, contentType, headers, curl: this.curlCommand };
-      } catch (err) {
-        const elapsed = Math.round(performance.now() - t0);
-        this.response = { status: 0, ok: false, body: String(err), elapsed, size: 0, contentType: '', headers: [], curl: this.curlCommand };
-      } finally {
-        this.sending = false;
-      }
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Agent Sandboxes page — sandbox branch (expand / collapse + HTMX lazy-load)
-// ---------------------------------------------------------------------------
-
-/**
- * Per-sandbox state for the Agent Sandboxes (worktrees) page.
- *
- * Handles:
- *   - toggle open/close with CSS transition
- *   - lazy HTMX detail load on first open (fires `detail-load` on the panel once)
- *   - single-sandbox delete via DELETE /api/worktrees/{slug}
- */
-function envSandbox(slug) {
-  return {
-    slug,
-    open: false,
-    loaded: false,
-
-    toggle() {
-      this.open = !this.open;
-      if (this.open && !this.loaded) {
-        this.loaded = true;
-        const detail = document.getElementById('env-detail-' + this.slug);
-        if (detail) htmx.trigger(detail, 'detail-load');
-      }
-    },
-
-    async deleteWorktree(s) {
-      if (!confirm(`Remove sandbox "${s}"?\n\nThis deletes the working directory. The branch is kept.`)) return;
-      const resp = await fetch(`/api/worktrees/${s}`, { method: 'DELETE' });
-      if (resp.ok) {
-        this.$el.style.opacity = '0';
-        this.$el.style.transform = 'translateX(-6px)';
-        setTimeout(() => location.reload(), 300);
-      } else {
-        const data = await resp.json().catch(() => ({}));
-        alert(`Failed to remove sandbox: ${data.detail ?? resp.status}`);
-      }
-    },
-  };
-}
-
-/* ═══════════════════════════════════════════════════════════════
-   agentsPage(batches, ghBaseUrl)
-   Alpine component for the /agents listing page.
-   Handles client-side filtering, sorting, and view toggling of
-   the run history.  Live agents are server-rendered by Jinja.
-   ═══════════════════════════════════════════════════════════════ */
-function agentsPage(batches, ghBaseUrl) {
-  return {
-    batches,
-    GH_BASE_URL: ghBaseUrl,
-
-    // Filter / sort / view state
-    activeStatus: '',
-    activeRole:   '',
-    searchQ:      '',
-    sortBy:       'newest',
-    view:         'cards',
-    openBatches:  {},   // batch_id → bool; undefined = open by default
-
-    init() {
-      // Open all batches by default.
-      this.batches.forEach(b => { this.openBatches[b.batch_id] = true; });
-    },
-
-    clearFilters() {
-      this.activeStatus = '';
-      this.activeRole   = '';
-      this.searchQ      = '';
-    },
-
-    toggleBatch(id) {
-      this.openBatches[id] = !(this.openBatches[id] !== false);
-    },
-
-    /** Format a kebab-case role string as Title Case words. */
-    formatRole(role) {
-      return (role || '').split('-').map(w => w.charAt(0).toUpperCase() + w.slice(1)).join(' ');
-    },
-
-    /** Human-readable label for spawn_mode values. */
-    formatSpawnMode(mode) {
-      const labels = { wave: 'Wave', chain: 'Chain', manual: 'Manual' };
-      return labels[mode] ?? mode ?? '';
-    },
-
-    /** CSS modifier class for spawn_mode chip. */
-    spawnModeClass(mode) {
-      return mode ? `run-spawn-mode run-spawn-mode--${mode}` : '';
-    },
-
-    /** Per-batch success rate string, e.g. "83%". */
-    batchSuccessRate(runs) {
-      if (!runs.length) return '—';
-      const done = runs.filter(r => r.status === 'done').length;
-      return Math.round(done / runs.length * 100) + '%';
-    },
-
-    /** Match a single run against current filters. */
-    matchRun(run) {
-      if (this.activeStatus && run.status !== this.activeStatus) return false;
-      if (this.activeRole   && run.role   !== this.activeRole)   return false;
-      if (this.searchQ) {
-        const q = this.searchQ.toLowerCase();
-        const hay = [
-          run.id, run.role, run.branch, run.batch_id,
-          run.issue_number ? '#' + run.issue_number : '',
-          run.pr_number    ? '#' + run.pr_number    : '',
-          run.spawn_mode   ?? '',
-        ].join(' ').toLowerCase();
-        if (!hay.includes(q)) return false;
-      }
-      return true;
-    },
-
-    /** Return filtered + sorted batches (only batches with ≥1 matching run). */
-    filteredBatches() {
-      let result = this.batches.map(batch => {
-        let runs = batch.runs.filter(r => this.matchRun(r));
-        if (this.sortBy === 'oldest')   runs = [...runs].reverse();
-        if (this.sortBy === 'issue')    runs = [...runs].sort((a, b) => (a.issue_number || 0) - (b.issue_number || 0));
-        if (this.sortBy === 'duration') runs = [...runs].sort((a, b) => (b.duration_s || 0) - (a.duration_s || 0));
-        return { ...batch, runs };
-      }).filter(b => b.runs.length > 0);
-
-      if (this.sortBy === 'oldest') result = [...result].reverse();
-      return result;
-    },
-
-    /** Total matching runs across all filtered batches. */
-    filteredTotal() {
-      return this.filteredBatches().reduce((n, b) => n + b.runs.length, 0);
-    },
-
-    /** Produce a status summary for a batch's runs [{status, count}, …]. */
-    batchStatusSummary(runs) {
-      const counts = {};
-      runs.forEach(r => { counts[r.status] = (counts[r.status] || 0) + 1; });
-      return Object.entries(counts).map(([status, count]) => ({ status, count }));
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// DAG visualisation — dependency graph with D3 force-directed layout
-// ---------------------------------------------------------------------------
-
-/**
- * Drives the Dependency DAG page.  Reads node/edge data from a JSON
- * <script type="application/json"> tag injected by Jinja2, then delegates
- * all rendering to D3.  Alpine manages the thin UI state layer:
- * filter, search, selected-node sidebar, and zoom controls.
- *
- * D3 is loaded from CDN in {% block head %} of dag.html — this function
- * guards against D3 not being present.
- */
-function dagVisualization() {
-  const PALETTE = [
-    '#3b82f6', '#6366f1', '#14b8a6', '#22c55e',
-    '#f97316', '#ef4444', '#a855f7', '#06b6d4', '#eab308', '#ec4899',
-  ];
-  const DEFAULT_COLOR = '#6b7280';
-  const NODE_RADIUS   = 20;
-
-  return {
-    // ── State ──────────────────────────────────────────────────────────────
-    rawNodes: [],
-    rawEdges: [],
-    phaseLabels: [],
-    phaseColors: {},
-    ghBaseUrl: '',
-
-    filter: 'all',
-    search: '',
-    selectedNode: null,
-    refreshing: false,
-
-    // D3 internals — not reactive (managed directly)
-    _sim: null,
-    _zoom: null,
-    _svg: null,
-    _container: null,
-    _linkGroup: null,
-    _nodeGroup: null,
-
-    // ── Computed ───────────────────────────────────────────────────────────
-    get filteredNodes() {
-      let nodes = this.rawNodes;
-      if (this.filter !== 'all') {
-        nodes = nodes.filter(n => Array.isArray(n.labels) && n.labels.includes(this.filter));
-      }
-      if (this.search.trim()) {
-        const q = this.search.trim().toLowerCase().replace(/^#/, '');
-        nodes = nodes.filter(n =>
-          String(n.number).includes(q) ||
-          (n.title || '').toLowerCase().includes(q)
-        );
-      }
-      return nodes;
-    },
-
-    get filteredEdges() {
-      const ids = new Set(this.filteredNodes.map(n => n.number));
-      return this.rawEdges.filter(([s, t]) => ids.has(s) && ids.has(t));
-    },
-
-    get sidebarOpen() {
-      return this.selectedNode !== null;
-    },
-
-    // ── Lifecycle ──────────────────────────────────────────────────────────
-    init() {
-      // Read injected JSON
-      const dataEl = document.getElementById('dag-data');
-      if (dataEl) {
-        try {
-          const parsed = JSON.parse(dataEl.textContent);
-          this.rawNodes    = parsed.nodes  || [];
-          this.rawEdges    = parsed.edges  || [];
-          this.phaseLabels = parsed.phase_labels || [];
-          this.ghBaseUrl   = parsed.gh_base_url  || '';
-        } catch (_) { /* data missing — graph stays empty */ }
-      }
-
-      // Build phase → colour map
-      this.phaseColors = Object.fromEntries(
-        this.phaseLabels.map((lbl, i) => [lbl, PALETTE[i % PALETTE.length]])
-      );
-
-      // Boot D3 once DOM is ready
-      this.$nextTick(() => this._initD3());
-    },
-
-    // ── D3 bootstrap ──────────────────────────────────────────────────────
-    _initD3() {
-      if (typeof d3 === 'undefined') return;
-
-      this._svg = d3.select('#dag-svg');
-
-      this._zoom = d3.zoom()
-        .scaleExtent([0.15, 5])
-        .on('zoom', (ev) => {
-          this._container.attr('transform', ev.transform);
-        });
-
-      this._svg.call(this._zoom);
-
-      this._container  = this._svg.append('g').attr('class', 'dag-container');
-      this._linkGroup  = this._container.append('g').attr('class', 'dag-links');
-      this._nodeGroup  = this._container.append('g').attr('class', 'dag-nodes');
-
-      // Arrow marker
-      this._svg.append('defs').append('marker')
-        .attr('id', 'dag-arrow')
-        .attr('viewBox', '0 0 10 10')
-        .attr('refX', 26)
-        .attr('refY', 5)
-        .attr('markerWidth', 6)
-        .attr('markerHeight', 6)
-        .attr('orient', 'auto-start-reverse')
-        .append('path')
-        .attr('d', 'M 0 0 L 10 5 L 0 10 z')
-        .attr('fill', '#6b7280');
-
-      this._render();
-
-      // Re-render whenever filter or search changes
-      this.$watch('filter', () => this._render());
-      this.$watch('search', () => this._render());
-    },
-
-    // ── D3 render ─────────────────────────────────────────────────────────
-    _nodeColor(node) {
-      if (!Array.isArray(node.labels)) return DEFAULT_COLOR;
-      for (const lbl of node.labels) {
-        if (this.phaseColors[lbl]) return this.phaseColors[lbl];
-      }
-      return DEFAULT_COLOR;
-    },
-
-    _render() {
-      if (!this._svg || typeof d3 === 'undefined') return;
-
-      const nodes = this.filteredNodes;
-      const edges = this.filteredEdges;
-      const self  = this;
-
-      if (this._sim) this._sim.stop();
-
-      const svgEl = document.getElementById('dag-svg');
-      const W = svgEl ? (svgEl.clientWidth  || 900) : 900;
-      const H = svgEl ? (svgEl.clientHeight || 520) : 520;
-
-      const nodeData = nodes.map(n => ({ ...n, id: String(n.number) }));
-      const edgeData = edges.map(([s, t]) => ({
-        source: String(s),
-        target: String(t),
-      }));
-
-      this._sim = d3.forceSimulation(nodeData)
-        .force('link', d3.forceLink(edgeData).id(d => d.id).distance(110).strength(0.5))
-        .force('charge', d3.forceManyBody().strength(-380))
-        .force('center', d3.forceCenter(W / 2, H / 2))
-        .force('collision', d3.forceCollide(NODE_RADIUS + 8));
-
-      // ── Links ────────────────────────────────────────────────────────────
-      const link = this._linkGroup
-        .selectAll('line')
-        .data(edgeData, d => `${d.source}-${d.target}`)
-        .join(
-          enter => enter.append('line')
-            .attr('stroke', '#6b7280')
-            .attr('stroke-width', 1.5)
-            .attr('marker-end', 'url(#dag-arrow)')
-            .attr('opacity', 0.65),
-          update => update,
-          exit   => exit.remove(),
-        );
-
-      // ── Nodes ────────────────────────────────────────────────────────────
-      const tooltip = document.getElementById('dag-tooltip');
-
-      const node = this._nodeGroup
-        .selectAll('g.dag-node')
-        .data(nodeData, d => d.id)
-        .join(
-          enter => {
-            const g = enter.append('g')
-              .attr('class', 'dag-node')
-              .style('cursor', 'pointer')
-              .call(d3.drag()
-                .on('start', (ev, d) => {
-                  if (!ev.active) self._sim.alphaTarget(0.3).restart();
-                  d.fx = d.x;
-                  d.fy = d.y;
-                })
-                .on('drag', (ev, d) => {
-                  d.fx = ev.x;
-                  d.fy = ev.y;
-                })
-                .on('end', (ev, d) => {
-                  if (!ev.active) self._sim.alphaTarget(0);
-                  d.fx = null;
-                  d.fy = null;
-                })
-              );
-
-            g.append('circle').attr('r', NODE_RADIUS);
-
-            g.append('text')
-              .attr('text-anchor', 'middle')
-              .attr('dominant-baseline', 'central')
-              .attr('font-size', '9px')
-              .attr('font-family', 'monospace')
-              .attr('fill', '#fff')
-              .attr('pointer-events', 'none')
-              .text(d => `#${d.number}`);
-
-            return g;
-          },
-          update => update,
-          exit   => exit.remove(),
-        );
-
-      // Apply live styles to all nodes
-      this._nodeGroup.selectAll('g.dag-node').each(function(d) {
-        const g      = d3.select(this);
-        const isClosed = String(d.state || '').toUpperCase() === 'CLOSED';
-        const isMatch  = self.search.trim()
-          ? String(d.number).includes(self.search.replace(/^#/, '')) ||
-            (d.title || '').toLowerCase().includes(self.search.toLowerCase())
-          : true;
-
-        g.select('circle')
-          .attr('fill',         self._nodeColor(d))
-          .attr('opacity',      isClosed ? 0.35 : isMatch ? 1.0 : 0.25)
-          .attr('stroke',       d.has_wip ? '#22c55e' : 'none')
-          .attr('stroke-width', d.has_wip ? 3 : 0);
-      });
-
-      // Click → select node (show sidebar)
-      this._nodeGroup.selectAll('g.dag-node')
-        .on('click', (_ev, d) => {
-          self.selectNode(d);
-        })
-        .on('mouseover', (ev, d) => {
-          if (!tooltip) return;
-          const blocking = d.blocking_count || 0;
-          tooltip.innerHTML =
-            `<span class="dag-tooltip-num">#${d.number}</span>` +
-            `<span class="dag-tooltip-title">${d.title || ''}</span>` +
-            `<div class="dag-tooltip-meta">` +
-            (d.has_wip ? `<span class="dag-tooltip-badge wip">WIP</span>` : '') +
-            (blocking > 0 ? `<span class="dag-tooltip-badge blocking">blocks ${blocking}</span>` : '') +
-            `<span class="dag-tooltip-badge">depth ${d.depth || 0}</span>` +
-            `</div>`;
-          tooltip.style.display = 'block';
-        })
-        .on('mousemove', (ev) => {
-          if (!tooltip) return;
-          const wrapper = document.querySelector('.dag-canvas-wrapper');
-          const rect    = wrapper ? wrapper.getBoundingClientRect() : { left: 0, top: 0 };
-          tooltip.style.left = (ev.clientX - rect.left + 14) + 'px';
-          tooltip.style.top  = (ev.clientY - rect.top  + 14) + 'px';
-        })
-        .on('mouseout', () => {
-          if (tooltip) tooltip.style.display = 'none';
-        });
-
-      // Tick
-      this._sim.on('tick', () => {
-        link
-          .attr('x1', d => d.source.x)
-          .attr('y1', d => d.source.y)
-          .attr('x2', d => d.target.x)
-          .attr('y2', d => d.target.y);
-
-        this._nodeGroup.selectAll('g.dag-node')
-          .attr('transform', d => `translate(${d.x},${d.y})`);
-      });
-    },
-
-    // ── Interaction helpers ────────────────────────────────────────────────
-    selectNode(nodeOrNum) {
-      if (typeof nodeOrNum === 'number' || typeof nodeOrNum === 'string') {
-        const num = parseInt(String(nodeOrNum).replace('#', ''), 10);
-        nodeOrNum = this.rawNodes.find(n => n.number === num) || null;
-      }
-      this.selectedNode = nodeOrNum || null;
-    },
-
-    closeNode() {
-      this.selectedNode = null;
-    },
-
-    setFilter(val) {
-      this.filter = val;
-    },
-
-    zoomIn() {
-      if (!this._svg || !this._zoom) return;
-      this._svg.transition().duration(250).call(this._zoom.scaleBy, 1.4);
-    },
-
-    zoomOut() {
-      if (!this._svg || !this._zoom) return;
-      this._svg.transition().duration(250).call(this._zoom.scaleBy, 1 / 1.4);
-    },
-
-    resetView() {
-      if (!this._svg || !this._zoom) return;
-      this._svg.transition().duration(300).call(this._zoom.transform, d3.zoomIdentity);
-      this._render();
-    },
-
-    async refreshData() {
-      this.refreshing = true;
-      try {
-        const r = await fetch('/api/dag');
-        if (!r.ok) throw new Error('HTTP ' + r.status);
-        const data       = await r.json();
-        this.rawNodes    = data.nodes || [];
-        this.rawEdges    = data.edges || [];
-        this._render();
-      } catch (_) { /* silently fail — stale data stays */ }
-      finally { this.refreshing = false; }
-    },
-
-    // ── Sidebar dep navigation ─────────────────────────────────────────────
-    nodeByNum(num) {
-      return this.rawNodes.find(n => n.number === num) || null;
-    },
-
-    depTitle(num) {
-      const n = this.nodeByNum(num);
-      return n ? n.title : `#${num}`;
-    },
-  };
-}
+"use strict";(()=>{var g=(t=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(t,{get:(e,s)=>(typeof require<"u"?require:e)[s]}):t)(function(t){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+t+'" is not supported')});function y(){return{projects:[],activeProject:null,async load(){try{let t=await fetch("/api/config");if(!t.ok)return;let e=await t.json();this.projects=e.projects||[],this.activeProject=e.active_project||null}catch{}},async switchProject(t){if(!(!t||t===this.activeProject))try{(await fetch("/api/config/switch-project",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({project_name:t})})).ok&&window.location.reload()}catch{}}}}function w(t){return{state:t,connected:!1,boardLoading:!1,_es:null,GH_BASE_URL:"",killModal:{show:!1,agentId:null,issueNumber:null,killing:!1},connect(){this.GH_BASE_URL=this.$el.dataset.ghBaseUrl??"",this._es=new EventSource("/events"),this._es.onopen=()=>{this.connected=!0},this._es.onmessage=e=>{try{this.state=JSON.parse(e.data),this.connected=!0}catch{}},this._es.onerror=()=>{this.connected=!1}},openKillModal(e,s){this.killModal={show:!0,agentId:e,issueNumber:s??null,killing:!1}},closeKillModal(){this.killModal.show=!1},async confirmKill(){let e=this.killModal.agentId;if(e){this.killModal.killing=!0;try{await fetch(`/api/control/kill/${encodeURIComponent(e)}`,{method:"POST"})}catch(s){console.error("Kill failed",s)}finally{this.killModal={show:!1,agentId:null,issueNumber:null,killing:!1}}}},relativeTime(e){if(!e)return"\u2014";let s=Math.floor(Date.now()/1e3-e);return s<5?"just now":s<60?s+"s ago":s<3600?Math.floor(s/60)+"m ago":Math.floor(s/3600)+"h ago"},agentForIssue(e){return!this.state||!this.state.agents?null:this.state.agents.find(s=>s.issue_number===e)||null},totalMsgCount(){return!this.state||!this.state.agents?0:this.state.agents.reduce((e,s)=>e+(s.message_count||0),0)}}}function _(){return{expanded:!1,transcript:[],transcriptLoading:!1,transcriptError:null,_prevMsgCount:0,toggle(t){this.expanded=!this.expanded,this.expanded&&this.transcript.length===0&&this.fetchTranscript(t)},async fetchTranscript(t){this.transcriptLoading=!0,this.transcriptError=null;try{let e=await fetch(`/api/agents/${encodeURIComponent(t)}/transcript`);if(!e.ok)throw new Error(`HTTP ${e.status}`);let s=await e.json();this.transcript=Array.isArray(s)?s:s.messages||[],this._scrollFeed()}catch(e){this.transcriptError=e.message||"Failed to load transcript"}finally{this.transcriptLoading=!1}},checkRefresh(t,e){let s=e||0;this.expanded&&s!==this._prevMsgCount&&this.fetchTranscript(t.id),this._prevMsgCount=s},_scrollFeed(){this.$nextTick(()=>{let t=this.$el.querySelector(".transcript-feed");t&&(t.scrollTop=t.scrollHeight)})},formatRole(t){return(t||"").split("-").map(e=>e.charAt(0).toUpperCase()+e.slice(1)).join(" ")},isActive(t){return t==="implementing"||t==="reviewing"},truncate(t,e){return t?t.length>e?t.slice(0,e)+"\u2026":t:""},parseArch(t){if(!t)return{emoji:"\u{1F916}",displayName:"",archetype:"",archetypeKey:"default",skillDomains:[],figureId:null};let e={steve_jobs:{name:"Steve Jobs",emoji:"\u{1F52E}",archetype:"The Visionary",key:"visionary"},satya_nadella:{name:"Satya Nadella",emoji:"\u{1F3DB}\uFE0F",archetype:"The Architect",key:"architect"},jeff_bezos:{name:"Jeff Bezos",emoji:"\u2699\uFE0F",archetype:"The Operator",key:"operator"},werner_vogels:{name:"Werner Vogels",emoji:"\u{1F3DB}\uFE0F",archetype:"The Architect",key:"architect"},linus_torvalds:{name:"Linus Torvalds",emoji:"\u26A1",archetype:"The Hacker",key:"hacker"},margaret_hamilton:{name:"Margaret Hamilton",emoji:"\u{1F6E1}\uFE0F",archetype:"The Guardian",key:"guardian"},bjarne_stroustrup:{name:"Bjarne Stroustrup",emoji:"\u26A1",archetype:"The Hacker",key:"hacker"},martin_fowler:{name:"Martin Fowler",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},kent_beck:{name:"Kent Beck",emoji:"\u{1F527}",archetype:"The Pragmatist",key:"pragmatist"},yann_lecun:{name:"Yann LeCun",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},andrej_karpathy:{name:"Andrej Karpathy",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},turing:{name:"Alan Turing",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},hopper:{name:"Grace Hopper",emoji:"\u{1F9D1}\u200D\u{1F3EB}",archetype:"The Mentor",key:"mentor"},dijkstra:{name:"Edsger Dijkstra",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},knuth:{name:"Donald Knuth",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},ritchie:{name:"Dennis Ritchie",emoji:"\u26A1",archetype:"The Hacker",key:"hacker"},guido_van_rossum:{name:"Guido van Rossum",emoji:"\u{1F527}",archetype:"The Pragmatist",key:"pragmatist"},mccarthy:{name:"John McCarthy",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},lovelace:{name:"Ada Lovelace",emoji:"\u{1F52E}",archetype:"The Visionary",key:"visionary"},von_neumann:{name:"John von Neumann",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},shannon:{name:"Claude Shannon",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},feynman:{name:"Richard Feynman",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},einstein:{name:"Albert Einstein",emoji:"\u{1F52E}",archetype:"The Visionary",key:"visionary"},hamming:{name:"Richard Hamming",emoji:"\u{1F4DA}",archetype:"The Scholar",key:"scholar"},bruce_schneier:{name:"Bruce Schneier",emoji:"\u{1F6E1}\uFE0F",archetype:"The Guardian",key:"guardian"}},s={the_visionary:{emoji:"\u{1F52E}",name:"The Visionary",key:"visionary"},the_architect:{emoji:"\u{1F3DB}\uFE0F",name:"The Architect",key:"architect"},the_hacker:{emoji:"\u26A1",name:"The Hacker",key:"hacker"},the_guardian:{emoji:"\u{1F6E1}\uFE0F",name:"The Guardian",key:"guardian"},the_scholar:{emoji:"\u{1F4DA}",name:"The Scholar",key:"scholar"},the_mentor:{emoji:"\u{1F9D1}\u200D\u{1F3EB}",name:"The Mentor",key:"mentor"},the_operator:{emoji:"\u2699\uFE0F",name:"The Operator",key:"operator"},the_pragmatist:{emoji:"\u{1F527}",name:"The Pragmatist",key:"pragmatist"}},a=t.split(":").map(d=>d.trim()).filter(Boolean),i=a[0],r=e[i];if(r)return{figureId:i,displayName:r.name,emoji:r.emoji,archetype:r.archetype,archetypeKey:r.key,skillDomains:a.slice(1)};let o=s[i];return o?{figureId:null,displayName:o.name,emoji:o.emoji,archetype:o.name,archetypeKey:o.key,skillDomains:a.slice(1)}:{figureId:null,displayName:a.join(" \xB7 "),emoji:"\u{1F916}",archetype:"",archetypeKey:"default",skillDomains:a}}}}function b(t,e,s){return{current:t,labels:e,pinned:s,open:!1,async selectLabel(a){this.open=!1,this.$dispatch("phase-switching");try{let i=await fetch("/api/control/active-label",{method:"PUT",headers:{"Content-Type":"application/json"},body:JSON.stringify({label:a})});if(i.ok){let r=await i.json();this.current=r.label,this.pinned=r.pinned,window.location.reload()}}catch{}},async selectAuto(){this.open=!1,this.$dispatch("phase-switching");try{let a=await fetch("/api/control/active-label",{method:"DELETE"});if(a.ok){let i=await a.json();this.current=i.label,this.pinned=!1,window.location.reload()}}catch{}}}}function v(t){return{paused:t??!1,async toggle(){let e=this.paused?"/api/control/resume":"/api/control/pause";try{let s=await fetch(e,{method:"POST"});if(s.ok){let a=await s.json();this.paused=a.paused}}catch{}}}}function T(){return{sweeping:!1,async sweep(){if(confirm("Delete all stale agent branches and clear orphan agent:wip labels?")){this.sweeping=!0;try{let t=await fetch("/api/control/sweep",{method:"POST"}),e=await t.json();if(!t.ok){alert(`Sweep failed: ${e.detail??t.status}`);return}let s=e.deleted_branches?.length??0,a=e.cleared_wip_labels?.length??0,i=e.errors?.length??0,r=[`Swept: ${s} branch${s!==1?"es":""} deleted, ${a} label${a!==1?"s":""} cleared`];i>0&&r.push(`(${i} error${i!==1?"s":""} \u2014 check logs)`),alert(r.join(" "))}catch(t){alert(`Network error: ${t.message}`)}finally{this.sweeping=!1}}}}}function S(){return{launching:!1,waveResult:null,waveError:null,init(){},unclaimedCount(){let t=this.state?.board_issues;return t?t.filter(e=>!e.claimed).length:0},claimedCount(){let t=this.state?.board_issues;return t?t.filter(e=>e.claimed).length:0},async startWave(){this.launching=!0,this.waveResult=null,this.waveError=null;try{let t=await fetch("/api/control/spawn-wave",{method:"POST"}),e=await t.json();t.ok?this.waveResult=e:this.waveError=e.detail??`HTTP ${t.status}`}catch(t){this.waveError=`Network error: ${t.message}`}finally{this.launching=!1}}}}function k(t){return{recommendation:t??null,dismissed:!1,applying:!1,applyError:null,async apply(){this.applying=!0,this.applyError=null;try{let e=await fetch("/api/intelligence/scaling-advice/apply",{method:"POST"});if(e.ok)this.dismissed=!0;else{let s=await e.json().catch(()=>({}));this.applyError=s.detail||`HTTP ${e.status}`}}catch(e){this.applyError=e.message||"Network error"}finally{this.applying=!1}}}}function j(t){return{violations:t??[],closing:null,async closeViolation(e){this.closing=e;try{(await fetch(`/api/intelligence/pr-violations/${e}/close`,{method:"POST"})).ok&&(this.violations=this.violations.filter(a=>a.pr_number!==e))}catch{}finally{this.closing=null}}}}function x(t){return{confirming:!1,clearing:!1,cleared:!1,error:null,async clear(e){this.confirming=!1,this.clearing=!0,this.error=null;try{let s=await fetch(`/api/intelligence/stale-claims/${e}/clear`,{method:"POST"});if(s.ok)this.cleared=!0;else{let a=await s.json().catch(()=>({}));this.error=a.detail||`HTTP ${s.status}`}}catch(s){this.error=s.message||"Network error"}finally{this.clearing=!1}}}}function C(t){return{analysisHtml:"",analyzing:!1,analyzeError:null,async analyze(){this.analyzing=!0,this.analyzeError=null;try{let e=await fetch(`/api/analyze/issue/${t}/partial`,{method:"POST"});this.analysisHtml=await e.text()}catch(e){this.analyzeError=e.message}finally{this.analyzing=!1}}}}function E(){return{mode:"single",selectedIssue:null,selectedRole:"python-developer",searchQ:"",waveRole:"python-developer",brainDump:"",labelPrefix:"",loading:!1,result:null,resultMode:null,formError:null,issues:[],activeLabel:"",unclaimedCount:0,init(){try{this.issues=JSON.parse(this.$el.dataset.issues||"[]")}catch{this.issues=[]}try{this.activeLabel=JSON.parse(this.$el.dataset.activeLabel||'""')}catch{this.activeLabel=""}this.unclaimedCount=parseInt(this.$el.dataset.unclaimedCount||"0",10)},get filteredIssues(){let t=this.searchQ.trim().toLowerCase();return t?this.issues.filter(e=>String(e.number).includes(t)||(e.title||"").toLowerCase().includes(t)||(e.phase_label||"").toLowerCase().includes(t)):this.issues},get liveUnclaimedCount(){return this.issues.filter(t=>!t.claimed).length},selectIssue(t){t.claimed||(this.selectedIssue=this.selectedIssue?.number===t.number?null:t)},selectRole(t){this.selectedRole=t},clearResult(){this.result=null,this.formError=null,this.resultMode=null},async submitSingle(){if(this.selectedIssue){this.clearResult(),this.loading=!0;try{let t=await fetch("/api/control/spawn",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({issue_number:this.selectedIssue.number,role:this.selectedRole})}),e=await t.json();t.ok?(this.result=e,this.resultMode="single",this.issues=this.issues.map(s=>s.number===this.selectedIssue.number?{...s,claimed:!0}:s)):this.formError=e.detail??`HTTP ${t.status}`}catch(t){this.formError=`Network error: ${t.message}`}finally{this.loading=!1}}},async submitWave(){this.clearResult(),this.loading=!0;try{let t=await fetch(`/api/control/spawn-wave?role=${encodeURIComponent(this.waveRole)}`,{method:"POST"}),e=await t.json();if(!t.ok)this.formError=e.detail??`HTTP ${t.status}`;else{this.result=e,this.resultMode="wave";let s=new Set((e.spawned||[]).map(a=>a.spawned));this.issues=this.issues.map(a=>s.has(a.number)?{...a,claimed:!0}:a)}}catch(t){this.formError=`Network error: ${t.message}`}finally{this.loading=!1}},async submitCoordinator(){if(this.brainDump.trim()){this.clearResult(),this.loading=!0;try{let t=await fetch("/api/control/spawn-coordinator",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({brain_dump:this.brainDump,label_prefix:this.labelPrefix})}),e=await t.json();t.ok?(this.result=e,this.resultMode="coordinator"):this.formError=e.detail??`HTTP ${t.status}`}catch(t){this.formError=`Network error: ${t.message}`}finally{this.loading=!1}}},onIssuesRefreshed(t){if(this.issues=t,this.selectedIssue){let e=this.issues.find(s=>s.number===this.selectedIssue.number);this.selectedIssue=e??null}}}}function $(t,e){return{batches:t,GH_BASE_URL:e,activeStatus:"",activeRole:"",searchQ:"",sortBy:"newest",view:"cards",openBatches:{},init(){this.batches.forEach(s=>{this.openBatches[s.batch_id]=!0})},clearFilters(){this.activeStatus="",this.activeRole="",this.searchQ=""},toggleBatch(s){this.openBatches[s]=this.openBatches[s]===!1},formatRole(s){return(s||"").split("-").map(a=>a.charAt(0).toUpperCase()+a.slice(1)).join(" ")},formatSpawnMode(s){return{wave:"Wave",chain:"Chain",manual:"Manual"}[s]??s??""},spawnModeClass(s){return s?`run-spawn-mode run-spawn-mode--${s}`:""},batchSuccessRate(s){if(!s.length)return"\u2014";let a=s.filter(i=>i.status==="done").length;return Math.round(a/s.length*100)+"%"},matchRun(s){if(this.activeStatus&&s.status!==this.activeStatus||this.activeRole&&s.role!==this.activeRole)return!1;if(this.searchQ){let a=this.searchQ.toLowerCase();if(![s.id,s.role,s.branch,s.batch_id,s.issue_number?"#"+s.issue_number:"",s.pr_number?"#"+s.pr_number:"",s.spawn_mode??""].join(" ").toLowerCase().includes(a))return!1}return!0},filteredBatches(){let s=this.batches.map(a=>{let i=a.runs.filter(r=>this.matchRun(r));return this.sortBy==="oldest"&&(i=[...i].reverse()),this.sortBy==="issue"&&(i=[...i].sort((r,o)=>(r.issue_number||0)-(o.issue_number||0))),this.sortBy==="duration"&&(i=[...i].sort((r,o)=>(o.duration_s||0)-(r.duration_s||0))),{...a,runs:i}}).filter(a=>a.runs.length>0);return this.sortBy==="oldest"&&(s=[...s].reverse()),s},filteredTotal(){return this.filteredBatches().reduce((s,a)=>s+a.runs.length,0)},batchStatusSummary(s){let a={};return s.forEach(i=>{a[i.status]=(a[i.status]||0)+1}),Object.entries(a).map(([i,r])=>({status:i,count:r}))}}}function P(){return{activeTab:"Trend",switchTab(t){this.activeTab=t,this.$nextTick(()=>{window.telemetry&&window.telemetry["render"+t]&&window.telemetry["render"+t]()})},init(){this.$nextTick(()=>{window.telemetry&&window.telemetry.renderTrend()})}}}function O(){let t=["#3b82f6","#6366f1","#14b8a6","#22c55e","#f97316","#ef4444","#a855f7","#06b6d4","#eab308","#ec4899"],e="#6b7280";return{rawNodes:[],rawEdges:[],phaseLabels:[],phaseColors:{},ghBaseUrl:"",filter:"all",search:"",selectedNode:null,refreshing:!1,_sim:null,_zoom:null,_svg:null,_container:null,_linkGroup:null,_nodeGroup:null,get filteredNodes(){let a=this.rawNodes;if(this.filter!=="all"&&(a=a.filter(i=>Array.isArray(i.labels)&&i.labels.includes(this.filter))),this.search.trim()){let i=this.search.trim().toLowerCase().replace(/^#/,"");a=a.filter(r=>String(r.number).includes(i)||(r.title||"").toLowerCase().includes(i))}return a},get filteredEdges(){let a=new Set(this.filteredNodes.map(i=>i.number));return this.rawEdges.filter(([i,r])=>a.has(i)&&a.has(r))},get sidebarOpen(){return this.selectedNode!==null},init(){let a=document.getElementById("dag-data");if(a)try{let i=JSON.parse(a.textContent);this.rawNodes=i.nodes||[],this.rawEdges=i.edges||[],this.phaseLabels=i.phase_labels||[],this.ghBaseUrl=i.gh_base_url||""}catch{}this.phaseColors=Object.fromEntries(this.phaseLabels.map((i,r)=>[i,t[r%t.length]])),this.$nextTick(()=>this._initD3())},_initD3(){typeof d3>"u"||(this._svg=d3.select("#dag-svg"),this._zoom=d3.zoom().scaleExtent([.15,5]).on("zoom",a=>{this._container.attr("transform",a.transform)}),this._svg.call(this._zoom),this._container=this._svg.append("g").attr("class","dag-container"),this._linkGroup=this._container.append("g").attr("class","dag-links"),this._nodeGroup=this._container.append("g").attr("class","dag-nodes"),this._svg.append("defs").append("marker").attr("id","dag-arrow").attr("viewBox","0 0 10 10").attr("refX",26).attr("refY",5).attr("markerWidth",6).attr("markerHeight",6).attr("orient","auto-start-reverse").append("path").attr("d","M 0 0 L 10 5 L 0 10 z").attr("fill","#6b7280"),this._render(),this.$watch("filter",()=>this._render()),this.$watch("search",()=>this._render()))},_nodeColor(a){if(!Array.isArray(a.labels))return e;for(let i of a.labels)if(this.phaseColors[i])return this.phaseColors[i];return e},_render(){if(!this._svg||typeof d3>"u")return;let a=this.filteredNodes,i=this.filteredEdges,r=this;this._sim&&this._sim.stop();let o=document.getElementById("dag-svg"),d=o&&o.clientWidth||900,p=o&&o.clientHeight||520,f=a.map(n=>({...n,id:String(n.number)})),m=i.map(([n,l])=>({source:String(n),target:String(l)}));this._sim=d3.forceSimulation(f).force("link",d3.forceLink(m).id(n=>n.id).distance(110).strength(.5)).force("charge",d3.forceManyBody().strength(-380)).force("center",d3.forceCenter(d/2,p/2)).force("collision",d3.forceCollide(28));let z=this._linkGroup.selectAll("line").data(m,n=>`${n.source}-${n.target}`).join(n=>n.append("line").attr("stroke","#6b7280").attr("stroke-width",1.5).attr("marker-end","url(#dag-arrow)").attr("opacity",.65),n=>n,n=>n.remove()),u=document.getElementById("dag-tooltip"),U=this._nodeGroup.selectAll("g.dag-node").data(f,n=>n.id).join(n=>{let l=n.append("g").attr("class","dag-node").style("cursor","pointer").call(d3.drag().on("start",(c,h)=>{c.active||r._sim.alphaTarget(.3).restart(),h.fx=h.x,h.fy=h.y}).on("drag",(c,h)=>{h.fx=c.x,h.fy=c.y}).on("end",(c,h)=>{c.active||r._sim.alphaTarget(0),h.fx=null,h.fy=null}));return l.append("circle").attr("r",20),l.append("text").attr("text-anchor","middle").attr("dominant-baseline","central").attr("font-size","9px").attr("font-family","monospace").attr("fill","#fff").attr("pointer-events","none").text(c=>`#${c.number}`),l},n=>n,n=>n.remove());this._nodeGroup.selectAll("g.dag-node").each(function(n){let l=d3.select(this),c=String(n.state||"").toUpperCase()==="CLOSED",h=r.search.trim()?String(n.number).includes(r.search.replace(/^#/,""))||(n.title||"").toLowerCase().includes(r.search.toLowerCase()):!0;l.select("circle").attr("fill",r._nodeColor(n)).attr("opacity",c?.35:h?1:.25).attr("stroke",n.has_wip?"#22c55e":"none").attr("stroke-width",n.has_wip?3:0)}),this._nodeGroup.selectAll("g.dag-node").on("click",(n,l)=>{r.selectNode(l)}).on("mouseover",(n,l)=>{if(!u)return;let c=l.blocking_count||0;u.innerHTML=`<span class="dag-tooltip-num">#${l.number}</span><span class="dag-tooltip-title">${l.title||""}</span><div class="dag-tooltip-meta">`+(l.has_wip?'<span class="dag-tooltip-badge wip">WIP</span>':"")+(c>0?`<span class="dag-tooltip-badge blocking">blocks ${c}</span>`:"")+`<span class="dag-tooltip-badge">depth ${l.depth||0}</span></div>`,u.style.display="block"}).on("mousemove",n=>{if(!u)return;let l=document.querySelector(".dag-canvas-wrapper"),c=l?l.getBoundingClientRect():{left:0,top:0};u.style.left=n.clientX-c.left+14+"px",u.style.top=n.clientY-c.top+14+"px"}).on("mouseout",()=>{u&&(u.style.display="none")}),this._sim.on("tick",()=>{z.attr("x1",n=>n.source.x).attr("y1",n=>n.source.y).attr("x2",n=>n.target.x).attr("y2",n=>n.target.y),this._nodeGroup.selectAll("g.dag-node").attr("transform",n=>`translate(${n.x},${n.y})`)})},selectNode(a){if(typeof a=="number"||typeof a=="string"){let i=parseInt(String(a).replace("#",""),10);a=this.rawNodes.find(r=>r.number===i)||null}this.selectedNode=a||null},closeNode(){this.selectedNode=null},setFilter(a){this.filter=a},zoomIn(){!this._svg||!this._zoom||this._svg.transition().duration(250).call(this._zoom.scaleBy,1.4)},zoomOut(){!this._svg||!this._zoom||this._svg.transition().duration(250).call(this._zoom.scaleBy,1/1.4)},resetView(){!this._svg||!this._zoom||(this._svg.transition().duration(300).call(this._zoom.transform,d3.zoomIdentity),this._render())},async refreshData(){this.refreshing=!0;try{let a=await fetch("/api/dag");if(!a.ok)throw new Error("HTTP "+a.status);let i=await a.json();this.rawNodes=i.nodes||[],this.rawEdges=i.edges||[],this._render()}catch{}finally{this.refreshing=!1}},nodeByNum(a){return this.rawNodes.find(i=>i.number===a)||null},depTitle(a){let i=this.nodeByNum(a);return i?i.title:`#${a}`}}}function L(){let t={max_eng_vps:1,max_qa_vps:1,pool_size_per_vp:4,active_labels_order:[],ab_mode:{enabled:!1,target_role:null,variant_a_file:null,variant_b_file:null},projects:[],active_project:null};return{config:{...t},lastSaved:null,activeTab:"allocation",newLabel:"",saving:!1,toast:{msg:"",cls:""},projectSaving:null,savedAt:null,_dragIdx:null,get totalAgents(){return(this.config.max_eng_vps+this.config.max_qa_vps)*this.config.pool_size_per_vp},get engSlots(){return this.config.max_eng_vps*this.config.pool_size_per_vp},get qaSlots(){return this.config.max_qa_vps*this.config.pool_size_per_vp},get engPct(){return this.totalAgents?Math.round(this.engSlots/this.totalAgents*100):50},get qaPct(){return this.totalAgents?Math.round(this.qaSlots/this.totalAgents*100):50},async init(){let e=this.$el.dataset.config;if(e)try{this.config={...t,...JSON.parse(e)}}catch{}try{let s=await fetch("/api/config");if(!s.ok)throw new Error("HTTP "+s.status);this.config={...t,...await s.json()}}catch(s){this._showToast("Failed to load config: "+s.message,"err")}this.lastSaved=JSON.parse(JSON.stringify(this.config))},setTab(e){this.activeTab=e},addLabel(){let e=this.newLabel.trim();if(e){if(this.config.active_labels_order.includes(e)){this._showToast("Label already in list.","err");return}this.config.active_labels_order.push(e),this.newLabel=""}},removeLabel(e){this.config.active_labels_order.splice(e,1)},onDragStart(e,s){this._dragIdx=s,e.dataTransfer.effectAllowed="move",e.dataTransfer.setData("text/plain",String(s))},onDragOver(e){e.currentTarget.classList.add("drag-over")},onDragLeave(e){e.currentTarget.classList.remove("drag-over")},onDrop(e,s){e.currentTarget.classList.remove("drag-over");let a=this._dragIdx;if(a===null||a===s)return;let i=this.config.active_labels_order,[r]=i.splice(a,1);i.splice(s,0,r),this._dragIdx=null},onDragEnd(){document.querySelectorAll(".cfg-label-item.drag-over").forEach(e=>{e.classList.remove("drag-over")}),this._dragIdx=null},ensureAbMode(){this.config.ab_mode||(this.config.ab_mode={enabled:!1,target_role:null,variant_a_file:null,variant_b_file:null})},async switchProject(e){this.projectSaving=e;try{let s=await fetch("/api/config/switch-project",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({project_name:e})});if(!s.ok){let i=await s.json().then(r=>r.detail||s.statusText).catch(()=>s.statusText);throw new Error(i)}let a=await s.json();this.config.active_project=a.active_project,this.lastSaved=JSON.parse(JSON.stringify(this.config)),this._showToast("\u2705 Switched to "+e,"ok")}catch(s){this._showToast("Switch failed: "+s.message,"err")}finally{this.projectSaving=null}},cancel(){this.lastSaved&&(this.config=JSON.parse(JSON.stringify(this.lastSaved)),this._showToast("Changes discarded.",""))},get isDirty(){return JSON.stringify(this.config)!==JSON.stringify(this.lastSaved)},async save(){this.saving=!0,this._showToast("","");try{let e=await fetch("/api/config",{method:"PUT",headers:{"Content-Type":"application/json"},body:JSON.stringify(this.config)});if(!e.ok){let s=await e.json().then(a=>a.detail||e.statusText).catch(()=>e.statusText);throw new Error(s)}this.config={...t,...await e.json()},this.lastSaved=JSON.parse(JSON.stringify(this.config)),this.savedAt=new Date().toLocaleTimeString([],{hour:"2-digit",minute:"2-digit"}),this._showToast("\u2705 Config saved.","ok")}catch(e){this._showToast("Save failed: "+e.message,"err")}finally{this.saving=!1}},_showToast(e,s){this.toast={msg:e,cls:s}}}}function N(t,e,s){return{slug:t,fileExists:e,personas:s,activeTab:"personas",selectedPersonaId:null,figure:"",atomOverrides:{},skills:[],copied:!1,init(){},loadInEditor(){window.dispatchEvent(new CustomEvent("role-load",{detail:{slug:this.slug}}))},get archPreview(){let a=[];this.figure&&a.push(this.figure);for(let i of this.skills)a.push(i);return a.length?`COGNITIVE_ARCH=${a.join(":")}`:"(select a figure or skill)"},applyPersona(a){let i=this.personas.find(r=>r.id===a);i&&(this.figure=a,this.atomOverrides=i.overrides?{...i.overrides}:{},this.skills=[],this.activeTab="composer")},async copyArchString(){try{await navigator.clipboard.writeText(this.archPreview)}catch{}this.copied=!0,setTimeout(()=>{this.copied=!1},1500)},resetComposer(){this.figure="",this.atomOverrides={},this.skills=[]}}}function R(){return{editor:null,_monacoBooted:!1,_monacoReady:!1,_pendingSlug:null,currentSlug:null,currentPath:null,status:"",statusClass:"",breadcrumb:'\u2190 click "View / Edit Prompt" to open a file',canSave:!1,canDiff:!1,diffVisible:!1,diffTitle:"",diffLines:[],diffCommitReady:!1,diffCommitting:!1,async loadRole(t){if(!this._monacoReady){this._pendingSlug=t,this._monacoBooted||(this._monacoBooted=!0,this.setStatus("Loading editor\u2026",""),this._injectMonaco());return}await this._doLoad(t)},_injectMonaco(){let e="https://cdn.jsdelivr.net/npm/monaco-editor@0.52.0/min/vs",s=document.createElement("script");s.src=`${e}/loader.js`,s.onload=()=>{g.config({paths:{vs:e}}),g(["vs/editor/editor.main"],()=>{if(this.editor)return;this.$refs.editorPlaceholder&&(this.$refs.editorPlaceholder.style.display="none"),this.editor=monaco.editor.create(this.$refs.editorContainer,{value:"",language:"markdown",theme:"vs-dark",automaticLayout:!1,minimap:{enabled:!1},wordWrap:"on",scrollBeyondLastLine:!1,readOnly:!0});let a=null;new ResizeObserver(()=>{clearTimeout(a),a=setTimeout(()=>this.editor&&this.editor.layout(),100)}).observe(this.$refs.editorContainer),this._monacoReady=!0,this._pendingSlug&&(this._doLoad(this._pendingSlug),this._pendingSlug=null)})},document.head.appendChild(s)},async _doLoad(t){if(!this.editor)return;let e=`.cursor/roles/${t}.md`;this.setStatus(`Loading ${e}\u2026`,"");try{let s=await fetch(`/api/roles/${encodeURIComponent(t)}`);if(!s.ok)throw new Error(`HTTP ${s.status}`);let a=await s.json();requestAnimationFrame(()=>{this.editor.setValue(a.content),this.editor.updateOptions({readOnly:!1}),this.editor.layout(),this.currentSlug=t,this.currentPath=e,this.breadcrumb=e,this.canSave=!0,this.canDiff=!0;let i=a.meta.last_commit_message||"(uncommitted)";this.setStatus(`${a.meta.line_count} lines \xB7 ${i}`,"ok")})}catch(s){this.setStatus(`Failed to load: ${s.message}`,"err")}},async saveRole(){if(!(!this.editor||!this.currentSlug)){this.canSave=!1,this.setStatus("Saving\u2026","");try{let t=await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}`,{method:"PUT",headers:{"Content-Type":"application/json"},body:JSON.stringify({content:this.editor.getValue()})});if(!t.ok)throw new Error(`HTTP ${t.status}`);let e=await t.json();this.canSave=!0;let s=e.diff?e.diff.split(`
+`).length:0;this.setStatus(`\u2705 Saved \u2014 ${s} line(s) changed`,"ok")}catch(t){this.canSave=!0,this.setStatus(`Save failed: ${t.message}`,"err")}}},async previewDiff(){if(!(!this.editor||!this.currentSlug)){this.diffTitle=`Diff \u2014 ${this.currentPath}`,this.diffLines=[],this.diffCommitReady=!1,this.diffVisible=!0;try{let t=await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}/diff`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({content:this.editor.getValue()})});if(!t.ok)throw new Error(`HTTP ${t.status}`);let e=await t.json();this.diffLines=this._parseDiff(e.diff),this.diffCommitReady=!0}catch(t){this.diffLines=[{cls:"cas-diff-empty",text:`Failed: ${t.message}`}]}}},_parseDiff(t){return!t||!t.trim()?[{cls:"cas-diff-empty",text:"No changes \u2014 content is identical to HEAD."}]:t.split(`
+`).map(e=>{let s="cas-diff-line";return e.startsWith("+")&&!e.startsWith("+++")?s+=" cas-diff-line--added":e.startsWith("-")&&!e.startsWith("---")?s+=" cas-diff-line--removed":e.startsWith("@@")&&(s+=" cas-diff-line--hunk"),{cls:s,text:e}})},async saveAndCommit(){if(!(!this.editor||!this.currentSlug)){this.diffCommitting=!0;try{let t=await fetch(`/api/roles/${encodeURIComponent(this.currentSlug)}/commit`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({content:this.editor.getValue()})});if(!t.ok)throw new Error(`HTTP ${t.status}`);let e=await t.json();this.diffVisible=!1,this.setStatus(`\u2705 Committed \u2014 SHA ${e.commit_sha.slice(0,8)} \xB7 ${e.message}`,"ok")}catch(t){this.setStatus(`Commit failed: ${t.message}`,"err")}finally{this.diffCommitting=!1}}},setStatus(t,e=""){this.status=t,this.statusClass=e}}}function I(){return{step:"input",funnelStage:0,text:"",labelPrefix:"",showOptions:!1,focused:!1,submitting:!1,errorMsg:"",result:{},loadingMsg:"",_loadingMsgs:[],_loadingTimer:null,init(){this._loadingMsgs=JSON.parse(this.$el.dataset.loadingMsgs||"[]"),this.loadingMsg=this._loadingMsgs[0]??""},get lineCount(){return this.text.split(`
+`).filter(t=>t.trim()).length},stageClass(t){return this.funnelStage>t?"bd-funnel-stage--done":this.funnelStage===t&&this.step!=="input"?"bd-funnel-stage--active bd-funnel-stage--pulse":t===0&&this.step==="input"?"bd-funnel-stage--active":""},autoGrow(t){t.style.height="auto",t.style.height=Math.min(t.scrollHeight,600)+"px"},async pasteClipboard(){try{let t=await navigator.clipboard.readText();this.text=(this.text?this.text+`
+`:"")+t,await this.$nextTick(),this.autoGrow(this.$refs.textarea)}catch{}},appendSeed(t){this.text=(this.text.trim()?this.text.trim()+`
+`:"")+t,this.$nextTick(()=>this.autoGrow(this.$refs.textarea))},_startLoadingAnimation(){let t=0;this.funnelStage=1,this.loadingMsg=this._loadingMsgs[0]??"",this._loadingTimer=setInterval(()=>{t=(t+1)%this._loadingMsgs.length,this.loadingMsg=this._loadingMsgs[t]??"",this.funnelStage=Math.min(t+1,5)},700)},_stopLoadingAnimation(t){this._loadingTimer&&(clearInterval(this._loadingTimer),this._loadingTimer=null),this.funnelStage=t},async submit(){let t=this.text.trim();if(t){this.submitting=!0,this.errorMsg="",this.step="loading",this._startLoadingAnimation();try{let e=await fetch("/api/control/spawn-coordinator",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({brain_dump:t,label_prefix:this.labelPrefix.trim()})});if(!e.ok){let a=await e.json().catch(()=>({}));throw new Error(a.detail||`HTTP ${e.status}`)}this.result=await e.json(),this._stopLoadingAnimation(5),this.step="done";let s=document.getElementById("bd-recent-runs");s&&typeof htmx<"u"&&htmx.trigger(s,"refresh")}catch(e){this._stopLoadingAnimation(0),this.errorMsg=e.message,this.step="input"}finally{this.submitting=!1}}},reset(){this._stopLoadingAnimation(0),this.step="input",this.funnelStage=0,this.text="",this.labelPrefix="",this.showOptions=!1,this.errorMsg="",this.result={}}}}function A(t,e,s){return{q:t||"",filterRole:e||"",filterStatus:s||"",sort:"mtime",applyFilters(){let a=this.q.toLowerCase();document.querySelectorAll("[data-transcript-row]").forEach(i=>{let r=(i.dataset.preview||"").toLowerCase(),o=i.dataset.role||"",d=i.dataset.status||"",p=!a||r.includes(a),f=!this.filterRole||o===this.filterRole,m=!this.filterStatus||d===this.filterStatus;i.style.display=p&&f&&m?"":"none"}),this.applySortDom()},applySortDom(){let a=document.getElementById("transcript-tbody");if(!a)return;let i=Array.from(a.querySelectorAll("[data-transcript-row]"));i.sort((r,o)=>{switch(this.sort){case"messages":return parseInt(o.dataset.messages)-parseInt(r.dataset.messages);case"words":return parseInt(o.dataset.words)-parseInt(r.dataset.words);case"subagents":return parseInt(o.dataset.subagents)-parseInt(r.dataset.subagents);default:return parseFloat(o.dataset.mtime)-parseFloat(r.dataset.mtime)}}),i.forEach(r=>a.appendChild(r))}}}function D(){return{search:"",highlight(t){if(!this.search)return this._esc(t);let e=this.search.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"),s=new RegExp(e,"gi");return this._esc(t).replace(s,a=>`<mark class="hl">${a}</mark>`)},_esc(t){return String(t).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;")}}}function M(){return{name:"",version:"",busy:!1,toast:{msg:"",cls:""},async doExport(){if(!(!this.name||!this.version)){this.busy=!0,this._toast("","");try{let t=await fetch("/api/templates/export",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({name:this.name,version:this.version})});if(!t.ok){let d=await t.json().catch(()=>({detail:t.statusText}));throw new Error(d.detail||t.statusText)}let e=await t.blob(),a=(t.headers.get("Content-Disposition")||"").match(/filename="([^"]+)"/),i=a?a[1]:"template.tar.gz",r=URL.createObjectURL(e),o=document.createElement("a");o.href=r,o.download=i,o.click(),URL.revokeObjectURL(r),this._toast("\u2705 Exported \u2014 page will refresh to show the new entry.","ok"),setTimeout(()=>location.reload(),1500)}catch(t){this._toast("Export failed: "+t.message,"err")}finally{this.busy=!1}}},_toast(t,e){this.toast={msg:t,cls:e}}}}function B(){return{file:null,targetRepo:"",busy:!1,toast:{msg:"",cls:""},conflicts:[],onFileChange(t){this.file=t.target.files[0]||null},async doImport(){if(!(!this.file||!this.targetRepo)){this.busy=!0,this.conflicts=[],this._toast("","");try{let t=new FormData;t.append("file",this.file);let e="/api/templates/import?target_repo="+encodeURIComponent(this.targetRepo),s=await fetch(e,{method:"POST",body:t});if(!s.ok){let i=await s.json().catch(()=>({detail:s.statusText}));throw new Error(i.detail||s.statusText)}let a=await s.json();this.conflicts=(a.conflicts||[]).filter(i=>i.exists),this._toast("\u2705 Imported "+a.extracted.length+" file(s) ("+a.name+" v"+a.version+").","ok")}catch(t){this._toast("Import failed: "+t.message,"err")}finally{this.busy=!1}}},_toast(t,e){this.toast={msg:t,cls:e}}}}function J(t){return{slug:t,open:!1,loaded:!1,toggle(){if(this.open=!this.open,this.open&&!this.loaded){this.loaded=!0;let e=document.getElementById("env-detail-"+this.slug);e&&htmx.trigger(e,"detail-load")}},async deleteWorktree(e){if(!confirm(`Remove sandbox "${e}"?
+
+This deletes the working directory. The branch is kept.`))return;let s=await fetch(`/api/worktrees/${e}`,{method:"DELETE"});if(s.ok)this.$el.style.opacity="0",this.$el.style.transform="translateX(-6px)",setTimeout(()=>location.reload(),300);else{let a=await s.json().catch(()=>({}));alert(`Failed to remove sandbox: ${a.detail??s.status}`)}}}}function H(){return{open:!1,tryOpen:!1,method:"GET",pathTemplate:"/",pathParamNames:[],pathParamValues:{},queryParams:[],queryValues:{},bodyFields:[],bodyJson:"",response:null,sending:!1,init(){this.method=this.$el.dataset.method||"GET",this.pathTemplate=this.$el.dataset.path||"/";try{this.queryParams=JSON.parse(this.$el.dataset.queryParams||"[]")}catch{this.queryParams=[]}for(let e of this.queryParams)this.queryValues[e.name]="";try{this.bodyFields=JSON.parse(this.$el.dataset.bodyFields||"[]")}catch{this.bodyFields=[]}if(this.bodyFields.length){let e={};for(let s of this.bodyFields)s.default!==null&&s.default!==void 0&&s.default!==""?e[s.name]=s.default:s.type==="integer"||s.type==="number"?e[s.name]=0:s.type==="boolean"?e[s.name]=!1:s.type&&s.type.startsWith("array")?e[s.name]=[]:e[s.name]="";this.bodyJson=JSON.stringify(e,null,2)}let t=this.pathTemplate.match(/\{(\w+)\}/g)||[];this.pathParamNames=t.map(e=>e.slice(1,-1));for(let e of this.pathParamNames)this.pathParamValues[e]=""},get resolvedPath(){let t=this.pathTemplate;for(let[s,a]of Object.entries(this.pathParamValues))t=t.replace(`{${s}}`,a?encodeURIComponent(a):`{${s}}`);let e=Object.entries(this.queryValues).filter(([,s])=>s!=="").map(([s,a])=>`${encodeURIComponent(s)}=${encodeURIComponent(a)}`);return e.length&&(t+="?"+e.join("&")),t},get curlCommand(){let t=window.location.origin,e=`curl -X ${this.method} '${t}${this.resolvedPath}'`;if(!["GET","HEAD"].includes(this.method)&&this.bodyJson.trim()){let a=this.bodyJson.replace(/'/g,"'\\''");e+=` \\
+  -H 'Content-Type: application/json' \\
+  -d '${a}'`}return e},formatSize(t){return t<1024?`${t} B`:t<1024*1024?`${(t/1024).toFixed(1)} KB`:`${(t/(1024*1024)).toFixed(1)} MB`},async send(){this.sending=!0,this.response=null;let t=performance.now();try{let e={method:this.method,headers:{}};!["GET","HEAD","DELETE"].includes(this.method)&&this.bodyJson.trim()&&(e.headers["Content-Type"]="application/json",e.body=this.bodyJson);let s=await fetch(this.resolvedPath,e),a=Math.round(performance.now()-t),i=await s.text(),r=[];s.headers.forEach((f,m)=>r.push({name:m,value:f}));let o=s.headers.get("content-type")||"",d=new TextEncoder().encode(i).length,p=i;try{p=JSON.stringify(JSON.parse(i),null,2)}catch{}this.response={status:s.status,ok:s.ok,body:p,elapsed:a,size:d,contentType:o,headers:r,curl:this.curlCommand}}catch(e){let s=Math.round(performance.now()-t);this.response={status:0,ok:!1,body:String(e),elapsed:s,size:0,contentType:"",headers:[],curl:this.curlCommand}}finally{this.sending=!1}}}}Object.assign(window,{projectSwitcher:y,pipelineDashboard:w,agentCard:_,phaseSwitcher:b,pipelineControl:v,sweepControl:T,waveControl:S,scalingAdvisor:k,prViolations:j,staleClaimCard:x,issueCard:C,agentsPage:$,missionControl:E,telemetryDash:P,dagVisualization:O,configPanel:L,roleDetail:N,rolesEditor:R,brainDump:I,transcriptBrowser:A,transcriptDetail:D,exportPanel:M,importPanel:B,envSandbox:J,apiEndpoint:H});})();
